Data.HList

Description

OOHaskell (C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

This module gathers the API that we need for OOP in Haskell. We basically select a certain configuration of the HList library, and we also import modules that are needed for mutable data and monads. Note on overlapping: Needed for the chosen model of labels. Other models can be used instead, but the chosen look better in types.

Documentation

module Data.HList.CommonMain

module Data.STRef

module Data.IORef

module Data.Typeable

module Control.Monad

module Control.Monad.ST

module Control.Monad.Fix

concrete :: MonadFix m => (a -> m a) -> a -> m a

(#) :: HasField l r v => r -> Label l -> v