Data.HList.FakePrelude

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Some very basic technology for faking dependent types in Haskell.

A heterogeneous apply operator

class Apply f a where

simpler/weaker version where type information only propagates forward with this one. app defined below, is more complicated / verbose to define, but it offers better type inference. Most uses have been converted to app, so there is not much that can be done with Apply.

Associated Types

type ApplyR f a :: *

Methods

apply :: f -> a -> ApplyR f a

Polymorphic functions are not first-class in haskell. One solution is to write an instance of ApplyAB for a data type that takes the place of the original function. In other words,

 data Fn = Fn
 instance ApplyAB Fn a b where applyAB Fn a = actual_fn a

Normally you would have been able to pass around the definition actual_fn.

Type inference / Local functional dependencies

Note that class ApplyAB has three parameters and no functional dependencies. Instances should be written in the style:

 instance (int ~ Int, double ~ Double) => ApplyAB Fn int double
  where applyAB _ = fromIntegral

rather than the more natural

 instance ApplyAB Fn Int Double

The first instance allows types to be inferred as if we had class ApplyAB a b c | a -> b c, while the second instance only matches if ghc already knows that it needs ApplyAB Fn Int Double. Additional explanation can be found in local functional dependencies

class ApplyAB f a b where

No constraints on result and argument types

Methods

applyAB :: f -> a -> b

Fun can be used instead of writing a new instance of ApplyAB. Refer to the definition/source for the the most concise explanation. A more wordy explanation is given below:

A type signature needs to be provided on Fun to make it work. Depending on the kind of the parameters to Fun, a number of different results happen.

ex1

A list of kind [* -> Constraint] produces those constraints on the argument type:

>>> :set -XDataKinds
>>> let plus1 = Fun (\x -> if x < 5 then x+1 else 5) :: Fun '[Num, Ord] '()
>>> :t applyAB plus1
applyAB plus1 :: (Num a, Ord a) => a -> a

Also note the use of '() to signal that the result type is the same as the argument type.

A single constraint can also be supplied:

>>> let succ1 = Fun succ :: Fun Enum '()
>>> :t applyAB succ1
applyAB succ1 :: Enum a => a -> a

>>> let just = Fun Just :: Fun '[] Maybe
>>> :t applyAB just
applyAB just :: a -> Maybe a

data Fun cxt getb

Constructors

Fun (forall a. FunCxt cxt a => a -> FunApp getb a)

data Fun' cxt geta

see Fun. The only difference here is that the argument type is calculated from the result type.

>>> let rd = Fun' read :: Fun' Read String
>>> :t applyAB rd
applyAB rd :: Read b => [Char] -> b

>>> let fromJust' = Fun' (\(Just a) -> a) :: Fun' '[] Maybe
>>> :t applyAB fromJust'
applyAB fromJust' :: Maybe b -> b

Note this use of Fun' means we don't have to get the b out of Maybe b,

Constructors

Fun' (forall b. FunCxt cxt b => FunApp geta b -> b)

type family FunApp fns a

type family FunCxt cxts a :: Constraint

Simple useful instances of Apply

data HPrint

print. An alternative implementation could be:

>>> let hPrint = Fun print :: Fun Show (IO ())

This produces:

>>> :t applyAB hPrint
applyAB hPrint :: Show a => a -> IO ()

Constructors

HPrint

data HRead

read

>>> applyAB HRead "5.0" :: Double
5.0

Constructors

HRead

data HShow

show

Constructors

HShow

data HComp g f

Compose two instances of ApplyAB

>>> applyAB (HComp HRead HShow) (5::Double) :: Double
5.0

Constructors

HComp g f
g . f

data Comp

app Comp (f,g) = g . f. Works like:

>>> applyAB Comp (succ, pred) 'a'
'a'

>>> applyAB Comp (toEnum :: Int -> Char, fromEnum) 10
10

Note that defaulting will sometimes give you the wrong thing

 used to work (with associated types calculating result/argument types)
 >>> applyAB Comp (fromEnum, toEnum) 'a'
 *** Exception: Prelude.Enum.().toEnum: bad argument

Constructors

Comp

newtype HSeq x

((a,b) -> f a >> b)

Constructors

HSeq x

data HFlip

flip

Constructors

HFlip

Proxy

see Data.HList.Proxy

data Label l

A special Proxy for record labels, polykinded

Constructors

Label

labelToProxy :: Label l -> Proxy l

class ShowLabel l where

Methods

showLabel :: Label l -> String

Booleans

GHC already lifts booleans, defined as

 data Bool = True | False

to types: Bool becomes kind and True and False (also denoted by 'True and 'False) become nullary type constructors.

The above line is equivalent to

 data HTrue
 data HFalse

 class HBool x
 instance HBool HTrue
 instance HBool HFalse

Value-level proxies

hTrue :: Proxy True

hFalse :: Proxy False

Conjunction

type family HAnd t1 t2 :: Bool

hAnd :: Proxy t1 -> Proxy t2 -> Proxy (HAnd t1 t2)

demote to values

Disjunction

type family HOr t1 t2 :: Bool

hOr :: Proxy t1 -> Proxy t2 -> Proxy (HOr t1 t2)

demote to values

Compare with the original code based on functional dependencies:

 class (HBool t, HBool t', HBool t'') => HOr t t' t'' | t t' -> t''
  where
   hOr :: t -> t' -> t''

 instance HOr HFalse HFalse HFalse
  where
   hOr _ _ = hFalse

 instance HOr HTrue HFalse HTrue
  where
   hOr _ _ = hTrue

 instance HOr HFalse HTrue HTrue
  where
   hOr _ _ = hTrue

 instance HOr HTrue HTrue HTrue
  where
   hOr _ _ = hTrue

class HCond t x y z | t x y -> z where

Methods

hCond :: Proxy t -> x -> y -> z

Boolean equivalence

type family HBoolEQ t1 t2 :: Bool

Naturals

data HNat

The data type to be lifted to the type level

Constructors

HZero

HSucc HNat

hZero :: Proxy HZero

hSucc :: Proxy (n :: HNat) -> Proxy (HSucc n)

hPred :: Proxy (HSucc n) -> Proxy n

class HNat2Integral n where

Methods

hNat2Integral :: Integral i => Proxy n -> i

type family HNatEq t1 t2 :: Bool

Equality on natural numbers (eventually to be subsumed by the universal polykinded HEq)

type family HLt x y :: Bool

Less than

hLt :: Proxy x -> Proxy y -> Proxy (HLt x y)

Maybies

We cannot use lifted Maybe since the latter are not populated

data HNothing

Constructors

HNothing

newtype HJust x

Constructors

HJust x

Polykinded Equality for types

class HEq x y b | x y -> b

We have to use Functional dependencies for now, for the sake of the generic equality.

hEq :: HEq x y b => x -> y -> Proxy b

Staged equality

Type-safe cast -- no longer need. We use a a ~ b

Error messages

class Fail x

A class without instances for explicit failure

module Data.HList.FakePrelude.Proxy