Data.HList.FakePrelude.Proxy

Description

Injection from algebraic kinds to *

Algebraic kinds like Nat are not populated and we can't use values of type Nat as function arguments. In contrast, we can use (undefined :: Proxy Z) as an argument, as a value proxy. data Proxy (tp :: k) :: *

re-exports Proxy if it exists

Documentation

data Proxy tp

proxy :: Proxy tp