Data.HList.GhcSyntax

Documentation

type :*: e l = HCons e l

(.*.) :: HExtend e l l' => e -> l -> l'

(.*.) -- Add a field to a record. Analagous to (++) for lists.

record .*. field1 .*. field2 -

type :=: l v = LVPair l v

(.=.) :: l -> v -> LVPair l v

(.=.) -- Create a value with the given label. Analagous to a data constructor such as Just, Left, or Right. Higher fixity than record-modification operations like (.*.), (.-.), etc. to support expression like the below w/o parentheses:

label1 .=. value1 .*. label2 .=. value2 .*. emptyRecord

(.!.) :: HasField l r v => r -> l -> v

(.!.) -- Lookup a value in a record, by its label. Analagous to (!!), the list indexing operation. Highest fixity, like (!!).

record1 .*. label1 .=. record2 .!. label1 .*. label2 .=. record2 .!. label2 -

(.-.) :: H2ProjectByLabels (HCons e HNil) t t1 r => Record t -> e -> Record r

(.-.) -- Remove a field from a record. At the same level as other record modification options (.*.). Analagous to (\) in lists.

record1 .-. label1

label1 .=. value1 .*. label2 .=. value2 .-. label2 .*. emptyRecord

label1 .=. value1 .-. label1 .*. label2 .=. value2 .*. emptyRecord

record1 .*. label1 .=. record2 .!. label1 .*. label2 .=. record2 .!. label2 .-. label1 -

(.@.) :: (HFind e l n, HUpdateAtHNat n (LVPair e v) t r, RecordLabels t l) => Record t -> LVPair e v -> Record r

(..) -- Update a field with a particular value. Another record-level operation, so it has the same fixity as (.*.) and (.-.). No real list analogue, since there is no Prelude defined update. record1 .. label1 .=. value1

record1 .. label1 .=. record2 .!. label1 record1 .*. label1 .=. record2 .!. label1 .*. label2 .=. record2 .!. label2 .. label1 .=. value2

(.^.) :: (HFind l ls n, HUpdateAtHNat n (LVPair l a) t l', HasField l t (Proxy a), RecordLabels t ls) => LVPair l a -> Record t -> Record l'

(.^.) -- This is a variation on updating (according to GhcRecord.hs), so use the same fixity as (.@.).

(.<.) :: (HFind e l n, HUpdateAtHNat n (LVPair e v) t r, HasField e t v, RecordLabels t l) => LVPair e v -> Record t -> Record r

(.<.) -- Another variation on update, so give it the same fixity as (.@.).

(.<++.) :: HLeftUnion r r' r'' => r -> r' -> r''

(.<++.) -- Similar to list append, so give this slightly lower fixity than (.*.), so we can write:

record1 .*. field1 .<++. record2 .*. field2

type :+: e l = HCons (Proxy e) l

(.+.) :: HExtend (Proxy e) l l' => e -> l -> l'

(.+.) -- Type-indexed rows append. Very similar to .*., so keep the same fixity.