Data.HList.HList

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Basic declarations for typeful heterogeneous lists.

Excuse the unstructured haddocks: while there are many declarations here some are alternative implementations should be grouped, and the definitions here are analgous to many list functions in the Prelude.

Heterogeneous type sequences

The easiest way to ensure that sequences can only be formed with Nil and Cons is to use GADTs The kind [*] is list kind (lists lifted to types)

data HList l where

Constructors

HNil :: HList []

HCons :: e -> HList l -> HList (e : l) infixr 2

Alternative representation

HNil' and HCons' are the older ADT-style. This has some advantages over the GADT:

lazy pattern matches are allowed

type inference is better if you want to directly pattern match see stackoverflow post here

data HNil'

Constructors

HNil'

data HCons' a b

Constructors

HCons' a b

class (UnPrime (Prime a) ~ a) => ConvHList a where

conversion between GADT (HList) and ADT (HNil' HCons') representations

Associated Types

type Prime a :: *

type UnPrime b :: [*]

Methods

prime :: HList a -> Prime a

unPrime :: Prime a -> HList a

Basic list functions

hHead :: HList (e : l) -> e

head

hTail :: HList (e : l) -> HList l

tail

type family HLength x :: HNat

Length

hLength :: HList l -> Proxy (HLength l)

Append

type family HAppendList l1 l2 :: [k]

hAppendList :: HList l1 -> HList l2 -> HList (HAppendList l1 l2)

the same as hAppend

Alternative append

append' :: [a] -> [a] -> [a]

hAppend' below is implemented using the same idea

hAppend' :: HFoldr FHCons v l r => HList l -> v -> r

Alternative implementation of hAppend. Demonstrates HFoldr

data FHCons

Constructors

FHCons

Historical append

The original HList code is included below. In both cases we had to program the algorithm twice, at the term and the type levels.

The class HAppend

class HAppend l l' l'' | l l' -> l''
 where
  hAppend :: l -> l' -> l''

The instance following the normal append

instance HList l => HAppend HNil l l
 where
  hAppend HNil l = l

instance (HList l, HAppend l l' l'')
      => HAppend (HCons x l) l' (HCons x l'')
 where
  hAppend (HCons x l) l' = HCons x (hAppend l l')

Reversing HLists

type family HRevApp l1 l2 :: [k]

hRevApp :: HList l1 -> HList l2 -> HList (HRevApp l1 l2)

hReverse :: HList l1 -> HList (HRevApp * l1 ([] *))

A nicer notation for lists

hEnd :: HList l -> HList l

Note:

x :: HList a

means: forall a. x :: HList a

hEnd x

means: exists a. x :: HList a

List termination

hBuild :: HBuild' [] r => r

Building lists

class HBuild' l r where

Methods

hBuild' :: HList l -> r

examples

The classes above allow the third (shortest) way to make a list (containing a,b,c) in this case

list = a `HCons` b `HCons` c `HCons` HNil
list = a .*. b .*. c .*. HNil
list = hEnd $ hBuild a b c

>>> let x = hBuild True in hEnd x
H[True]

>>> let x = hBuild True 'a' in hEnd x
H[True, 'a']

>>> let x = hBuild True 'a' "ok" in hEnd x
H[True, 'a', "ok"]

historical

the show instance has since changed, but these uses of 'hBuild'/'hEnd' still work

HList> let x = hBuild True in hEnd x
HCons True HNil

HList> let x = hBuild True 'a' in hEnd x
HCons True (HCons 'a' HNil)

HList> let x = hBuild True 'a' "ok" in hEnd x
HCons True (HCons 'a' (HCons "ok" HNil))

HList> hEnd (hBuild (Key 42) (Name "Angus") Cow (Price 75.5))
HCons (Key 42) (HCons (Name "Angus") (HCons Cow (HCons (Price 75.5) HNil)))

HList> hEnd (hBuild (Key 42) (Name "Angus") Cow (Price 75.5)) == angus
True

folds

foldr

Consume a heterogenous list. GADTs and type-classes mix well

class HFoldr f v l r where

Methods

hFoldr :: f -> v -> HList l -> r

class HScanr f z ls rs where

Methods

hScanr :: f -> z -> HList ls -> HList rs

class HFoldr1 f l r where

Methods

hFoldr1 :: f -> HList l -> r

foldl

class HFoldl f z xs r where

like foldl

>>> hFoldl (uncurry $ flip (:)) [] (1 `HCons` 2 `HCons` HNil)
[2,1]

Methods

hFoldl :: f -> z -> HList xs -> r

unfold

Produce a heterogenous list. Uses the more limited Apply instead of App since that's all that is needed for uses of this function downstream. Those could in principle be re-written.

hUnfold :: (Apply p s, HUnfold' p (ApplyR p s)) => p -> s -> HList (HUnfold p s)

type HUnfold p s = HUnfoldR p (ApplyR p s)

class HUnfold' p res where

Associated Types

type HUnfoldR p res :: [*]

Methods

hUnfold' :: p -> res -> HList (HUnfoldR p res)

replicate

class (HLength (HReplicateR n e) ~ n) => HReplicate n e where

Methods

hReplicate :: Proxy n -> e -> HList (HReplicateR n e)

type family HReplicateR n e :: [k]

would be associated with HReplicate except we want it to work with e of any kind, not just * that you can put into a HList. An "inverse" of HLength

concat

class HConcat a where

Like concat but for HLists of HLists.

Works in ghci... puzzling as what is different in doctest (it isn't -XExtendedDefaultRules)

hConcat $ hBuild (hBuild 1 2 3) (hBuild 'a' "abc")

H[1, 2, 3, a, "abc"]

Associated Types

type HConcatR a :: [*]

Methods

hConcat :: HList a -> HList (HConcatR a)

type family UnHList a :: [*]

traversing HLists

producing HList

map

It could be implemented with hFoldr, as we show further below

hMap :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux f as' bs') => f -> HList as' -> HList bs'

hMap is written such that the length of the result list can be determined from the length of the argument list (and the other way around). Similarly, the type of the elements of the list is propagated in both directions too.

Excuse the ugly types printed. Unfortunately ghc (still?) shows types like '[a,b] using the actual constructors involved (':) a ((':) b '[]) (or even worse when the kind variables are printed).

>>> :set -XNoMonomorphismRestriction
>>> let xs = 1 .*. 'c' .*. HNil
>>> :t hMap (HJust ()) xs
hMap (HJust ()) xs
  :: Num y => HList ((':) * (HJust y) ((':) * (HJust Char) ('[] *)))

These 4 examples show that the constraint on the length (2 in this cae) can be applied before or after the hMap. That inference is independent of the direction that type information is propagated for the individual elements.

>>> let asLen2 xs = xs `asTypeOf` (undefined :: HList '[a,b])

>>> let lr xs = asLen2 (applyAB (HMap HRead) xs)
>>> let ls xs = asLen2 (applyAB (HMap HShow) xs)
>>> let rl xs = applyAB (HMap HRead) (asLen2 xs)
>>> let sl xs = applyAB (HMap HShow) (asLen2 xs)

>>> :t lr
lr
  :: (Read y, Read y1) =>
     HList ((':) * String ((':) * String ('[] *)))
     -> HList ((':) * y ((':) * y1 ('[] *)))

>>> :t rl
rl
  :: (Read y, Read y1) =>
     HList ((':) * String ((':) * String ('[] *)))
     -> HList ((':) * y ((':) * y1 ('[] *)))

>>> :t ls
ls
  :: (Show y, Show y1) =>
     HList ((':) * y ((':) * y1 ('[] *)))
     -> HList ((':) * String ((':) * String ('[] *)))

>>> :t sl
sl
  :: (Show y, Show y1) =>
     HList ((':) * y ((':) * y1 ('[] *)))
     -> HList ((':) * String ((':) * String ('[] *)))

newtype HMap f

Constructors

HMap f

type HMapCxt f as bs as' bs' = (HMapAux f as' bs', as ~ HList as', bs ~ HList bs', SameLength as' bs')

class SameLength' es1 es2

Ensure two lists have the same length. We do case analysis on the first one (hence the type must be known to the type checker). In contrast, the second list may be a type variable.

class (SameLength' x y, SameLength' y x) => SameLength x y

symmetrical version of SameLength'. Written as a class instead of

type SameLength a b = (SameLength' a b, SameLength' b a)

since ghc expands type synonyms, but not classes (and it seems to have the same result)

class HMapAux f l r where

Methods

hMapAux :: SameLength l r => f -> HList l -> HList r

alternative implementation

currently broken

newtype MapCar f

Constructors

MapCar f

hMapMapCar :: HFoldr (MapCar f) (HList []) l l' => f -> HList l -> l'

Same as hMap only a different implementation.

appEndo . mconcat . map Endo

hComposeList :: HFoldr Comp (a -> a) l (t -> a) => HList l -> t -> a

>>> let xs = length .*. (+1) .*. (*2) .*. HNil
>>> hComposeList xs "abc"
8

sequence

class (Applicative m, SameLength a b) => HSequence m a b | a -> b, m b -> a where

A heterogeneous version of

sequenceA :: (Applicative m) => [m a] -> m [a]

Only now we operate on heterogeneous lists, where different elements may have different types a. In the argument list of monadic values (m a_i), although a_i may differ, the monad m must be the same for all elements. That's why we needed Data.HList.TypeCastGeneric2 (currently (~)). The typechecker will complain if we attempt to use hSequence on a HList of monadic values with different monads.

The hSequence problem was posed by Matthias Fischmann in his message on the Haskell-Cafe list on Oct 8, 2006

http://www.haskell.org/pipermail/haskell-cafe/2006-October/018708.html

http://www.haskell.org/pipermail/haskell-cafe/2006-October/018784.html

Maybe

>>> hSequence $ Just (1 :: Integer) `HCons` (Just 'c') `HCons` HNil
Just H[1, 'c']

>>> hSequence $  return 1 `HCons` Just  'c' `HCons` HNil
Just H[1, 'c']

List

>>> hSequence $ [1] `HCons` ['c'] `HCons` HNil
[H[1, 'c']]

Methods

hSequence :: HList a -> m (HList b)

newtype LiftA2 f

Constructors

LiftA2 f

alternative implementation

producing homogenous lists

map (no sequencing)

This one we implement via hFoldr

newtype Mapcar f

Constructors

Mapcar f

type HMapOut f l e = HFoldr (Mapcar f) [e] l [e]

hMapOut :: forall f e l. HMapOut f l e => f -> HList l -> [e]

mapM

hMapM :: (Monad m, HMapOut f l (m e)) => f -> HList l -> [m e]

mapM :: forall b m a. (Monad m) => (a -> m b) -> [a] -> m [b]

Likewise for mapM_.

See hSequence if the result list should also be heterogenous.

hMapM_ :: (Monad m, HMapOut f l (m ())) => f -> HList l -> m ()

GHC doesn't like its own type. hMapM_ :: forall m a f e. (Monad m, HMapOut f a (m e)) => f -> a -> m () Without explicit type signature, it's Ok. Sigh. Anyway, Hugs does insist on a better type. So we restrict as follows:

Type-level equality for lists (HEq)

Ensure a list to contain HNats only

type family HNats l :: [HNat]

We do so constructively, converting the HList whose elements are Proxy HNat to [HNat]. The latter kind is unpopulated and is present only at the type level.

hNats :: HList l -> Proxy (HNats l)

Membership tests

class HMember e1 l b | e1 l -> b

Check to see if an HList contains an element with a given type This is a type-level only test

class HMember' b0 e1 l b | b0 e1 l -> b

type family HMemberP pred e1 l :: Bool

type family HMemberP' pred e1 l pb :: Bool

hMember :: HMember e l b => Proxy e -> Proxy l -> Proxy b

Another type-level membership test

class HMemberM e1 l r | e1 l -> r

Check to see if an element e occurs in a list l If not, return 'Nothing If the element does occur, return 'Just l1 where l1 is a type-level list without e

class HMemberM1 b e1 l r | b e1 l -> r

class HMemberM2 b e1 l r | b e1 l -> r

Staged equality for lists

removed. use Typeable instead

Find an element in a set based on HEq

class HFind e l n | e l -> n

It is a pure type-level operation

class HFind' b e l n | b e l -> n

Membership test based on type equality

class HTMember e l b | e l -> b

could be an associated type if HEq had one

hTMember :: HTMember e l b => e -> HList l -> Proxy b

Intersection based on HTMember

class HTIntersect l1 l2 l3 | l1 l2 -> l3 where

Methods

hTIntersect :: HList l1 -> HList l2 -> HList l3

Like intersect

class HTIntersectBool b h t l1 l2 | b h t l1 -> l2 where

Methods

hTIntersectBool :: Proxy b -> h -> HList t -> HList l1 -> HList l2

Turn a heterogeneous list into a homogeneous one

class HList2List l e where

Same as hMapOut Id

Methods

hList2List :: HList l -> [e]

With HMaybe

Turn list in a list of justs

class ToHJust l l' | l -> l', l' -> l where

the same as map Just

>>> toHJust (2 .*. 'a' .*. HNil)
H[HJust 2, HJust 'a']

>>> toHJust2 (2 .*. 'a' .*. HNil)
H[HJust 2, HJust 'a']

Methods

toHJust :: HList l -> HList l'

toHJust2 :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux (HJust ()) as' bs') => HList as' -> HList bs'

alternative implementation. The Apply instance is in Data.HList.FakePrelude. A longer type could be inferred. toHJust2 :: (HMap' (HJust ()) a b) => HList a -> HList b

Extract justs from list of maybes

class FromHJust l where

Associated Types

type FromHJustR l :: [*]

Methods

fromHJust :: HList l -> HList (FromHJustR l)

alternative implementation

fromHJust2 :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux HFromJust as' bs') => HList as' -> HList bs'

A longer type could be inferred. fromHJust2 :: (HMap' HFromJust a b) => HList a -> HList b

data HFromJust

Constructors

HFromJust

Annotated lists

data HAddTag t

Constructors

HAddTag t

data HRmTag

Constructors

HRmTag

hAddTag :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux (HAddTag t) as' bs') => t -> HList as' -> HList bs'

hRmTag :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux HRmTag as' bs') => HList as' -> HList bs'

hFlag :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux (HAddTag (Proxy Bool True)) as' bs') => HList as' -> HList bs'

Annotate list with a type-level Boolean hFlag :: HMap' (HAddTag (Proxy True)) l r => HList l -> HList r

Splitting by HTrue and HFalse

class HSplit l where

Analogus to Data.List.partition snd

>>> hSplit $ (2,hTrue) .*. (3,hTrue) .*. (1,hFalse) .*. HNil
(H[2, 3],H[1])

it might make more sense to instead have LVPair Bool e instead of (e, Proxy Bool) since the former has the same runtime representation as e

Associated Types

type HSplitT l :: [*]

type HSplitF l :: [*]

Methods

hSplit :: HList l -> (HList (HSplitT l), HList (HSplitF l))