Data.HList.HListPrelude

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Declarations for various classes and functions that apply for the whole range of heterogeneous collections (HList, TIP, records, etc).

Documentation

class HExtend e l where

Associated Types

type HExtendR e l

Methods

(.*.) :: e -> l -> HExtendR e l infixr 2

emptyProxy :: Proxy [*] ([] *)

similar to emptyRecord, emptyTIP, emptyHList (actually called HNil), except emptyProxy is the rightmost argument to .*.

class SubType l l'

class HAppend l1 l2 where

Methods

hAppend :: l1 -> l2 -> HAppendR l1 l2

type family HAppendR l1 l2 :: k

poly-kinded, but hAppend only works in cases where the kind variable k is *

class HOccurs e l where

Methods

hOccurs :: l -> e

class HOccursNot e l

class HProject l l' where

Methods

hProject :: l -> l'

class HType2HNat e l n | e l -> n

Map a type (key) to a natural (index) within the collection This is a purely type-level computation

class HTypes2HNats es l ns | es l -> ns

class HDeleteMany e l l' | e l -> l' where

Delete all elements with the type-level key e from the collection l. Since the key is type-level, it is represented by a Proxy. (polykinded)

Methods

hDeleteMany :: Proxy e -> l -> l'

class HDeleteAtLabel r l v v' | l v -> v' where

Methods

hDeleteAtLabel :: Label l -> r v -> r v'

class SameLengths `[x, y, xy]` => HUnzip r x y xy | x y -> xy, xy -> x y where

unzip

Methods

hUnzip :: r xy -> (r x, r y)

class HUnzip r x y xy => HZip r x y xy where

zip. Variant supports hUnzip, but not hZip (hZipVariant returns a Maybe)

Methods

hZip :: r x -> r y -> r xy