Data.HList.HOccurs

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Result-type-driven operations on typeful heterogeneous lists.

Documentation

class HOccurrence e1 l l' | e1 l -> l' where

Methods

hOccurrence :: Proxy e1 -> HList l -> HList l'

class HOccurrence' b e1 l l' | b e1 l -> l' where

Methods

hOccurrence' :: Proxy b -> Proxy e1 -> HList l -> HList l'

class HOccursMany e l where

Methods

hOccursMany :: HList l -> [e]

class HOccursMany' e l where

Methods

hOccursMany' :: HList l -> [e]

hOccursMany1 :: forall e l l'. (HOccurrence e l (e : l'), HOccursMany e l') => HList l -> (e, [e])

hOccursFst :: forall e l l'. HOccurrence e l (e : l') => HList l -> e

data TypeNotFound e

class HOccurs' e l where

Methods

hOccurs' :: HList l -> e

hOccursRest :: (HOccurs l (r v), HDeleteAtLabel * r l v v') => r v -> (l, r v')

lookup a value in the collection (TIP usually) and return the TIP with that element deleted. Used to implement tipyTuple.

hOccursOpt :: forall e l l'. (HOccurrence e l l', HOccursOpt' e l') => HList l -> Maybe e

class HOccursOpt' e l where

Methods

hOccursOpt' :: HList l -> Maybe e

data TypeFound e

class HOccursNot' b e l

Illustration of typical test scenarios

Retrieve the Breed of an animal.

ghci-or-hugs> hOccurs myAnimal :: Breed
Cow

Normal hOccurs requires specification of the result type even if the result type is determined by the fact that we are faced with a singleton list.

ghci-or-hugs> hOccurs (HCons 1 HNil)

<interactive>:1:
    No instance for (HOccurs e1 (HCons e HNil))

However, hOccurs can be elaborated as improved as follows:

ghci-or-hugs> hLookup (HCons 1 HNil)
1