Data.HList.HSort

Description

Benchmarks for these functions can be found at http://code.haskell.org/~aavogt/HList-nodup/Run.html.

See Data-HList-CommonMain.html#v:hSort for the public interface.

Documentation

data HLeFn

the "standard" <= for types. Reuses HEqBy

Note that ghc-7.6 is missing instances for Symbol and Nat, so that sorting only works HNat (as used by Data.HList.Label3).

data HDown a

analogous to Down

data HNeq le

The HEqBy instances for HNeq HLeFn gives <

class HEqByFn le => HIsAscList le xs b | le xs -> b

HIsAscList le xs b is analogous to

b = all (\(x,y) -> x `le` y) (xs `zip` tail xs)

class (SameLength a b, HEqByFn le) => HSortBy le a b | le a -> b where

quick sort with a special case for sorted lists

Methods

hSortBy :: Proxy le -> HList a -> HList b

type HSort x y = HSortBy HLeFn x y

hSort :: HSort x y => HList x -> HList y

class HSortBy1 ok le a b | ok le a -> b where

Methods

hSortBy1 :: Proxy ok -> Proxy le -> HList a -> HList b

Merge Sort

class HEqByFn le => HMSortBy le a b | le a -> b where

HMSortBy is roughly a transcription of this merge sort

msort [] = []
msort [x] = [x]
msort [x,y] = hSort2 x y
msort xs = case splitAt (length xs `div` 2) xs of
             (a,b) -> msort a `merge` msort b

hSort2 x y
    | x <= y    = [x,y]
    | otherwise = [y,x]

merge (x : xs) (y : ys)
  | x > y     = y : merge (x : xs) ys
  | otherwise = x : merge xs (y : ys)

Methods

hMSortBy :: Proxy le -> HList a -> HList b

class HSort2 b x y ab | b x y -> ab where

Methods

hSort2 :: Proxy b -> x -> y -> HList ab

class HMerge le x y xy | le x y -> xy where

Methods

hMerge :: Proxy le -> HList x -> HList y -> HList xy

type HMerge1 b x y min max = (HCond b (HList x) (HList y) (HList min), HCond b (HList y) (HList x) (HList max))

hMerge1 :: (HCond t2 y x t, HCond t2 x y t1) => Proxy Bool t2 -> y -> x -> (t, t1)

Quick sort

class HQSortBy le a b | le a -> b where

HQSortBy is this algorithm

qsort (x : xs @ (_ : _)) = case partition (<= x) xs of
                 (le, gt) -> qsort le ++ x : qsort gt
qsort xs = xs

on random inputs that are not pathological (ie. not already sorted or reverse sorted) this turns out to be faster than HMSortBy, so it is used by default.

Methods

hQSortBy :: Proxy le -> HList a -> HList b

More efficient HRLabelSet / HLabelSet

class HEqByFn lt => HSetBy lt ps

Provided the labels involved have an appropriate instance of HEqByFn, it would be possible to use the following definitions:

type HRLabelSet = HSet
type HLabelSet  = HSet

class HSetBy (HNeq HLeFn) ps => HSet ps

class HIsSet ps b | ps -> b

>>> let xx = Proxy :: HIsSet [Label "x", Label "x"] b => Proxy b
>>> :t xx
xx :: Proxy 'False

>>> let xy = Proxy :: HIsSet [Label "x", Label "y"] b => Proxy b
>>> :t xy
xy :: Proxy 'True

class HEqByFn lt => HIsSetBy lt ps b | lt ps -> b

class HEqByFn le => HAscList le ps

HAscList le xs confirms that xs is in ascending order, and reports which element is duplicated otherwise.

class HEqByFn le => HAscList0 le ps ps0

class HEqByFn le => HAscList1 le b ps ps0

>>> import Data.HList.TypeEqO