Data.HList.Record

Description

The HList library

(C) 2004-2006, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Extensible records

The three-ish models of labels that go with this module;

Data.HList.Label3

Data.HList.MakeLabels

These used to work:

Data.HList.Label1

Data.HList.Label2

Data.HList.Label4

Data.HList.Label5

labels used for doctests

>>> let x = Label :: Label "x"
>>> let y = Label :: Label "y"
>>> let z = Label :: Label "z"

Records

Labels

Record types as label-value pairs, where label is purely phantom. Thus the run-time representation of a field is the same as that of its value, and the record, at run-time, is indistinguishable from the HList of field values. At run-time, all information about the labels is erased.

newtype LVPair l v

Field of label l with value type v Polykinded with respect to l: label may be a symbol, a nat, etc.

Constructors

LVPair

Fields

valueLVPair :: v

labelLVPair :: LVPair l v -> Label l

Label accessor

newLVPair :: Label l -> v -> LVPair l v

(.=.) :: Label l -> v -> LVPair l v

Create a value with the given label. Analagous to a data constructor such as Just, Left, or Right. Higher fixity than record-modification operations like (.*.), (.-.), etc. to support expression like the below w/o parentheses:

>>> x .=. "v1" .*. y .=. '2' .*. emptyRecord
Record{x="v1",y='2'}

(.-.) :: H2ProjectByLabels `[l]` r _r' r' => Record r -> Label l -> Record r'

Remove a field from a record. At the same level as other record modification options (.*.). Analagous to (\\) in lists.

 record1 .-. label1

 label1 .=. value1 .*.
 label2 .=. value2 .-.
 label2 .*.
 emptyRecord

 label1 .=. value1 .-.
 label1 .*.
 label2 .=. value2 .*.
 emptyRecord

 record1 .*. label1 .=. record2 .!. label1
         .*. label2 .=. record2 .!. label2
         .-. label1

Record

newtype Record r

Constructors

Record (HList r)

mkRecord :: HRLabelSet r => HList r -> Record r

Build a record

emptyRecord :: Record `[]`

Build an empty record

Getting Labels

type family RecordLabels r :: [k]

Construct the (phantom) list of labels of the record.

recordLabels :: Record r -> Proxy (RecordLabels r)

type family LabelsOf ls :: [*]

Making this ls::[*] and [k] breaks the MainGhcGeneric1.hs...

hLabels :: HList l -> Proxy (LabelsOf l)

Getting Values

class RecordValues r where

Construct the HList of values of the record.

Associated Types

type RecordValuesR r :: [*]

Methods

recordValues' :: HList r -> HList (RecordValuesR r)

recordValues :: RecordValues r => Record r -> HList (RecordValuesR r)

Operations

Show

A corresponding Show instance exists as

 show x = "Record {" ++ showComponents "" x ++ "}"

class ShowComponents l where

Methods

showComponents :: String -> HList l -> String

class ShowLabel l where

Methods

showLabel :: Label l -> String

Delete

hDeleteAtLabel label record

hDeleteAtLabel :: forall l t t1 t2. H2ProjectByLabels `[l]` t t1 t2 => Label l -> Record t -> Record t2

Lookup/update

Lens-based setters/getters are popular.

This is a provisional method to make a Lens (Record s) (Record t) a b, out of a Label x. Refer to examples/lens.hs for an example.

hLens :: (Functor f, HFind k l (RecordLabels k r) n, HUpdateAtHNat n (LVPair k l v) r, HasField k l (Record r) v1) => Label k l -> (v1 -> f v) -> Record r -> f (Record (HUpdateAtHNatR n (LVPair k l v) r))

Lookup

class HasField l r v | l r -> v where

This is a baseline implementation. We use a helper class, HasField, to abstract from the implementation.

Because hLookupByLabel is so frequent and important, we implement it separately, more efficiently. The algorithm is familiar assq, only the comparison operation is done at compile-time

Methods

hLookupByLabel :: Label l -> r -> v

class HasField' b l r v | b l r -> v where

Methods

hLookupByLabel' :: Proxy b -> Label l -> HList r -> v

(.!.) :: HasField l r v => r -> Label l -> v

Lookup a value in a record by its label. Analagous to (!!), the list indexing operation. Highest fixity, like (!!).

>>> :{
let record1 = x .=. 3 .*.
              y .=. 'y' .*.
              emptyRecord
:}

>>> record1 .!. x
3

>>> record1 .!. y
'y'

>>> :{
let r2 = y .=. record1 .!. x .*.
         z .=. record1 .!. y .*.
         emptyRecord
:}

>>> r2
Record{y=3,z='y'}

Note that labels made following Data.HList.Labelable allow using Control.Lens.^. instead.

Update

(.@.) :: (HFind k l (RecordLabels k r) n, HUpdateAtHNat n (LVPair k l v) r) => LVPair k l v -> Record r -> Record (HUpdateAtHNatR n (LVPair k l v) r)

Update a field with a particular value. Same fixity as (.*.) so that extensions and updates can be chained. There is no real list analogue, since there is no Prelude defined update.

 label1 .=. value1 .@. record1

hUpdateAtLabel :: forall r l n v. (HFind l (RecordLabels r) n, HUpdateAtHNat n (LVPair l v) r) => Label l -> v -> Record r -> Record (HUpdateAtHNatR n (LVPair l v) r)

hUpdateAtLabel label value record

type-preserving versions

(.<.) :: (HFind k l (RecordLabels k r) n, HUpdateAtHNat n (LVPair k l v) r, HasField k l (Record r) v) => LVPair k l v -> Record r -> Record (HUpdateAtHNatR n (LVPair k l v) r)

The same as .@., except type preserving. It has the same fixity as (.@.).

hTPupdateAtLabel :: (HFind k l (RecordLabels k r) n, HUpdateAtHNat n (LVPair k l v) r, HasField k l (Record r) v) => Label k l -> v -> Record r -> Record (HUpdateAtHNatR n (LVPair k l v) r)

A variation on hUpdateAtLabel: type-preserving update.

We could also say:

 hTPupdateAtLabel l v r = hUpdateAtLabel l v r `asTypeOf` r

Then we were taking a dependency on Haskell's type equivalence. This would also constrain the actual implementation of hUpdateAtLabel.

Rename Label

hRenameLabel :: (H2ProjectByLabels k (: k l ([] k)) t t1 t2, HasField k l (Record t) v, HRLabelSet (: * (LVPair k1 l1 v) t2)) => Label k l -> Label k1 l1 -> Record t -> Record (: * (LVPair k1 l1 v) t2)

Rename the label of record

>>> hRenameLabel x y (x .=. () .*. emptyRecord)
Record{y=()}

Projection

It is also an important operation: the basis of many deconstructors -- so we try to implement it efficiently.

hProjectByLabels :: (HRLabelSet a, H2ProjectByLabels ls t a b) => Proxy ls -> Record t -> Record a

hProjectByLabels ls r returns r with only the labels in ls remaining

hProjectByLabels2 :: (H2ProjectByLabels ls t t1 t2, HRLabelSet t1, HRLabelSet t2) => Proxy ls -> Record t -> (Record t1, Record t2)

See H2ProjectByLabels

Unions

Left

class HLeftUnion r r' r'' | r r' -> r'' where

Methods

hLeftUnion :: Record r -> Record r' -> Record r''

class HLeftUnionBool b r f r' | b r f -> r' where

Methods

hLeftUnionBool :: Proxy b -> Record r -> f -> Record r'

(.<++.) :: HLeftUnion r r' r'' => Record r -> Record r' -> Record r''

Similar to list append, so give this slightly lower fixity than (.*.), so we can write:

 field1 .=. value .*. record1 .<++. record2

Symmetric

Compute the symmetric union of two records r1 and r2 and return the pair of records injected into the union (ru1, ru2).

To be more precise, we compute the symmetric union type ru of two record types r1 and r2. The emphasis on types is important.

The two records (ru1,ru2) in the result of unionSR have the same type ru, but they are generally different values. Here the simple example: suppose

  r1 = (Label .=. True)  .*. emptyRecord
  r2 = (Label .=. False) .*. emptyRecord

Then unionSR r1 r2 will return (r1,r2). Both components of the result are different records of the same type.

To project from the union ru, use hProjectByLabels. It is possible to project from the union obtaining a record that was not used at all when creating the union.

We do assure however that if unionSR r1 r2 gave (r1u,r2u), then projecting r1u onto the type of r1 gives the value identical to r1. Ditto for r2.

class UnionSymRec r1 r2 ru | r1 r2 -> ru where

Methods

unionSR :: Record r1 -> Record r2 -> (Record ru, Record ru)

Reorder Labels

hRearrange :: (HLabelSet ls, HRearrange ls r (HList r')) => Proxy ls -> Record r -> Record r'

Rearranges a record by labels. Returns the record r, rearranged such that the labels are in the order given by ls. (recordLabels r) must be a permutation of ls.

Extension

hExtend, hAppend

(.*.) :: HExtend e l => e -> l -> HExtendR e l

Hints for type errors

data DuplicatedLabel l

Propery of a proper label set for a record: no duplication of labels

data ExtraField l

Constructors

ExtraField

data FieldNotFound l

Constructors

FieldNotFound

Unclassified

Probably internals, that may not be useful

class H2ProjectByLabels ls r rin rout | ls r -> rin rout where

Invariant:

 r === rin `disjoint-union` rout
 labels rin === ls
     where (rin,rout) = hProjectByLabels ls r

Methods

h2projectByLabels :: Proxy ls -> HList r -> (HList rin, HList rout)

class H2ProjectByLabels' b ls r rin rout | b ls r -> rin rout where

Methods

h2projectByLabels' :: Proxy b -> Proxy ls -> HList r -> (HList rin, HList rout)

class HLabelSet ls

Relation between HLabelSet and HRLabelSet

class HLabelSet' l1 l2 leq r

class HRLabelSet ps

class HRLabelSet' l1 l2 leq r

class HRearrange ls r r' where

Helper class for hRearrange

Methods

hRearrange2 :: Proxy ls -> HList r -> r'

class HRearrange' l ls rin rout r' where

Helper class 2 for hRearrange

Methods

hRearrange2' :: Proxy l -> Proxy ls -> HList rin -> HList rout -> r'

class UnionSymRec' b r1 f2 r2' ru | b r1 f2 r2' -> ru where

Methods

unionSR' :: Proxy b -> Record r1 -> f2 -> Record r2' -> (Record ru, Record ru)