Data.HList.TypeEqO

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Generic type equality predicate: The implementation based on overlapping instances The only place where overlapping instances are really used besides Label5

Documentation

class TupleType t b | t -> b

class IsKeyFN t flag | t -> flag

All our keywords must be registered