Data.HList.Variant

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Variants, i.e., labelled sums.

One approach to their implementation would be to consider both the favoured label and the corresponding value as dynamics upon variant construction. Since we are too lazy to programme some Typeable instances for non-ghc systems (NB: in GHC, Typeable is derivable), we rather model variants as (opaque) records with maybies for the values. Only one value will actually hold non-Nothing, as guaranteed by the constructor.

See VariantP.hs for a different approach to open sums.

Documentation

newtype Variant mr

Variant types on the basis of label-maybe pairs.

Constructors

Variant mr

data HMaybeF

Turn proxy sequence into sequence of Nothings

Constructors

HMaybeF

hMaybied :: (SameLength' * * bs' as', SameLength' * * as' bs', HMapAux HMaybeF as' bs') => HList as' -> HList bs'

mkVariant :: (HUpdateAtHNat n (Tagged k l (Maybe a)) r, SameLength' * * r as', SameLength' * * as' r, HMapAux HMaybeF as' r, HFind k l (RecordLabels k r) n, (~) [*] (HUpdateAtHNatR n (Tagged k l (Maybe a)) r) r) => Label k l -> a -> Record as' -> Variant (HList r)

Public constructor it seems we can blame hUpdateAtLabel (not HMap) for needing the asTypeOf?

unVariant :: HasField k l (Record r) v => Label k l -> Variant (HList r) -> v

Public destructor