Data.HList.FakePrelude

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Some very basic technology for faking dependent types in Haskell.

A heterogeneous apply operator

class Apply f a where

simpler/weaker version where type information only propagates forward with this one. applyAB defined below, is more complicated / verbose to define, but it offers better type inference. Most uses have been converted to applyAB, so there is not much that can be done with Apply.

Associated Types

type ApplyR f a :: *

Methods

apply :: f -> a -> ApplyR f a

Polymorphic functions are not first-class in haskell. An example of this is:

f op = (op (1 :: Double), op (1 :: Int))

RankNTypes

One solution is to enable `-XRankNTypes` and then write a type signature which might be `f :: (forall a. Num a => a -> a)`. This does not work in the context of HList, since we want to use functions that do not necessarily fall into the pattern of (forall a. c a => a -> a).

MultipleArguments

Another solution is to rewrite op to look like

f op1 op2 = (op1 (1:: Double), op2 (1 :: Int))

In some sense this approach works (see HZip), but the result is constrained to as many function applications as you are willing to write (ex. a function that works for records of six entries would look like hBuild f f f f f f).

Defunctionalization

Therefore the selected solution is to write an instance of ApplyAB for a data type that takes the place of the original function. In other words,

data Fn = Fn
instance ApplyAB Fn a b where applyAB Fn a = actual_fn a

Normally you would have been able to pass around the definition actual_fn.

Type inference / Local functional dependencies

Note that class ApplyAB has three parameters and no functional dependencies. Instances should be written in the style:

instance (int ~ Int, double ~ Double) => ApplyAB Fn int double
 where applyAB _ = fromIntegral

rather than the more natural

instance ApplyAB Fn Int Double

The first instance allows types to be inferred as if we had class ApplyAB a b c | a -> b c, while the second instance only matches if ghc already knows that it needs ApplyAB Fn Int Double. Since applyAB Fn :: Int -> Double has a monomorphic type, this trimmed down example does not really make sense because applyAB (fromIntegral :: Int -> Double) is exactly the same. Nontheless, the other uses of ApplyAB follow this pattern, and the benefits are seen when the type of applyAB Fn has at least one type variable.

Additional explanation can be found in local functional dependencies

AmbiguousTypes

Note that ghc only allows AllowAmbiguousTypes when a type signature is provided. Thus expressions such as:

data AddJust = AddJust
instance (y ~ Maybe x) => ApplyAB AddJust x y where
   applyAB _ x = Just x

twoJustsBad = hMap AddJust . hMap AddJust -- ambiguous type

Are not accepted without a type signature that references the intermediate "b":

twoJusts :: forall r a b c. (HMapCxt r AddJust a b, HMapCxt r AddJust b c) =>
       r a -> r c
twoJusts a = hMap AddJust (hMap AddJust a :: r b)

An apply class with functional dependencies

class ApplyAB' f a b | f a -> b, f b -> a

Or with equivalent type families

class (GetB f a ~ b, GetA f b ~ a) => ApplyAB' f a b

would not require an annotation for twoJusts. However, not all instances of ApplyAB will satisfy those functional dependencies, and thus the number of classes would proliferate. Furthermore, inference does not have to be in one direction only, as the example of HMap shows.

class ApplyAB f a b where

No constraints on result and argument types

Methods

applyAB :: f -> a -> b

Fun can be used instead of writing a new instance of ApplyAB. Refer to the definition/source for the the most concise explanation. A more wordy explanation is given below:

A type signature needs to be provided on Fun to make it work. Depending on the kind of the parameters to Fun, a number of different results happen.

ex1

A list of kind [* -> Constraint] produces those constraints on the argument type:

>>> :set -XDataKinds
>>> let plus1f x = if x < 5 then x+1 else 5
>>> let plus1 = Fun plus1f :: Fun '[Num, Ord] '()
>>> :t applyAB plus1
applyAB plus1 :: (Num b, Ord b) => b -> b

>>> let xs = [1 .. 8]
>>> map (applyAB plus1) xs == map plus1f xs
True

Also note the use of '() to signal that the result type is the same as the argument type.

A single constraint can also be supplied:

>>> let succ1 = Fun succ :: Fun Enum '()
>>> :t applyAB succ1
applyAB succ1 :: Enum b => b -> b

>>> let just = Fun Just :: Fun '[] Maybe
>>> :t applyAB just
applyAB just :: a -> Maybe a

data Fun cxt getb

Constructors

Fun (forall a. FunCxt cxt a => a -> FunApp getb a)

data Fun' cxt geta

see Fun. The only difference here is that the argument type is calculated from the result type.

>>> let rd = Fun' read :: Fun' Read String
>>> :t applyAB rd
applyAB rd :: Read b => [Char] -> b

>>> let fromJust' = Fun' (\(Just a) -> a) :: Fun' '[] Maybe
>>> :t applyAB fromJust'
applyAB fromJust' :: Maybe b -> b

Note this use of Fun' means we don't have to get the b out of Maybe b,

Constructors

Fun' (forall b. FunCxt cxt b => FunApp geta b -> b)

type family FunApp fns a

type family FunCxt cxts a :: Constraint

Simple useful instances of Apply

data HPrint

print. An alternative implementation could be:

>>> let hPrint = Fun print :: Fun Show (IO ())

This produces:

>>> :t applyAB hPrint
applyAB hPrint :: Show a => a -> IO ()

Constructors

HPrint

data HRead

read

>>> applyAB HRead "5.0" :: Double
5.0

Constructors

HRead

data HShow

show

Constructors

HShow

data HComp g f

Compose two instances of ApplyAB

>>> applyAB (HComp HRead HShow) (5::Double) :: Double
5.0

Constructors

HComp g f
g . f

data Comp

app Comp (f,g) = g . f. Works like:

>>> applyAB Comp (succ, pred) 'a'
'a'

>>> applyAB Comp (toEnum :: Int -> Char, fromEnum) 10
10

Note that defaulting will sometimes give you the wrong thing

used to work (with associated types calculating result/argument types)
>>> applyAB Comp (fromEnum, toEnum) 'a'
*** Exception: Prelude.Enum.().toEnum: bad argument

Constructors

Comp

newtype HSeq x

((a,b) -> f a >> b)

Constructors

HSeq x

data HFlip

flip

Constructors

HFlip

newtype HFmap f

fmap

Constructors

HFmap f

newtype LiftA2 f

liftA2

Constructors

LiftA2 f

data HUntag

untag

Constructors

HUntag

Proxy

see Data.Proxy

data Label l

A special Proxy for record labels, polykinded

Constructors

Label

labelToProxy :: Label l -> Proxy l

class ShowLabel l where

Methods

showLabel :: Label l -> String

Booleans

GHC already lifts booleans, defined as

data Bool = True | False

to types: Bool becomes kind and True and False (also denoted by 'True and 'False) become nullary type constructors.

The above line is equivalent to

data HTrue
data HFalse

class HBool x
instance HBool HTrue
instance HBool HFalse

Value-level proxies

hTrue :: Proxy True

hFalse :: Proxy False

Conjunction

type family HAnd t1 t2 :: Bool

hAnd :: Proxy t1 -> Proxy t2 -> Proxy (HAnd t1 t2)

demote to values

Disjunction

type family HOr t1 t2 :: Bool

hOr :: Proxy t1 -> Proxy t2 -> Proxy (HOr t1 t2)

demote to values

Compare with the original code based on functional dependencies:

class (HBool t, HBool t', HBool t'') => HOr t t' t'' | t t' -> t''
 where
  hOr :: t -> t' -> t''

instance HOr HFalse HFalse HFalse
 where
  hOr _ _ = hFalse

instance HOr HTrue HFalse HTrue
 where
  hOr _ _ = hTrue

instance HOr HFalse HTrue HTrue
 where
  hOr _ _ = hTrue

instance HOr HTrue HTrue HTrue
 where
  hOr _ _ = hTrue

type family HNot x :: Bool

class HNotFD b nb | b -> nb, nb -> b

as compared with HNot this version is injective

hNot :: HNotFD a notA => Proxy a -> Proxy notA

class HCond t x y z | t x y -> z where

Methods

hCond :: Proxy t -> x -> y -> z

Boolean equivalence

type family HBoolEQ t1 t2 :: Bool

Naturals

data HNat

The data type to be lifted to the type level

Constructors

HZero

HSucc HNat

hZero :: Proxy HZero

hSucc :: Proxy (n :: HNat) -> Proxy (HSucc n)

hPred :: Proxy (HSucc n) -> Proxy n

class HNat2Integral n where

Methods

hNat2Integral :: Integral i => Proxy n -> i

type family HNat2Nat n :: Nat

class HNats2Integrals ns where

Methods

hNats2Integrals :: Integral i => Proxy ns -> [i]

type family HNatEq t1 t2 :: Bool

Equality on natural numbers (eventually to be subsumed by the universal polykinded HEq)

type family HLt x y :: Bool

Less than

hLt :: Proxy x -> Proxy y -> Proxy (HLt x y)

type family HLe x y :: Bool

Less than or equal to

hLe :: Proxy x -> Proxy y -> Proxy (HLe x y)

type family HDiv2 x :: HNat

HDiv2 x behaves like x div 2

Maybies

We cannot use lifted Maybe since the latter are not populated

data HNothing

Constructors

HNothing

newtype HJust x

Constructors

HJust x

Polykinded Equality for types

class HEq x y b | x y -> b

We have to use Functional dependencies for now, for the sake of the generic equality.

type HEqK x y b = HEq (Proxy x) (Proxy y) b

Equality for types that may have different kinds. This definition allows operations on Record [Tagged "x" a, Tagged 2 b] to work as expected.

hEq :: HEq x y b => x -> y -> Proxy b

class HEqByFn f => HEqBy f x y b | f x y -> b

this class generalizes HEq by allowing the choice of f to allow equating only part of x and y

class HEqByFn f

Every instance of this class should have an instance of HEqBy

Arity

type Arity f n = (ArityFwd f n, ArityRev f n)

class ArityFwd f n | f -> n

calculate the number of arguments a function can take

class ArityRev f n

given the number of arguments a function can take, make sure the function type actually matches

Staged equality

Type-safe cast -- no longer need. We use a a ~ b

Cast

class HCast x y where

Named after cast, which behaves the same at runtime. One difference is that there is a HCast instance for every type, while Typeable instances can be missing sometimes.

Methods

hCast :: x -> Maybe y

class HCast1 b x y where

helper for HCast

Methods

hCast1 :: Proxy b -> x -> Maybe y

Error messages

class Fail x

A class without instances for explicit failure

Uses of fail

these could be replaced by `'("helpful message", l)`, but these look better.

data ExtraField l

data FieldNotFound l

type TypeablePolyK a = Typeable a

Constraining Lists

Length

class SameLength' es1 es2

Ensure two lists have the same length. We do case analysis on the first one (hence the type must be known to the type checker). In contrast, the second list may be a type variable.

class (SameLength' x y, SameLength' y x) => SameLength x y where

symmetrical version of SameLength'. Written as a class instead of

type SameLength a b = (SameLength' a b, SameLength' b a)

since ghc expands type synonyms, but not classes (and it seems to have the same result)

Minimal complete definition

Nothing

Methods

sameLength :: (r x `p` f (q y)) -> r x `p` f (q y)

SameLength x y => Equality (r x) (q y) (r x) (q y)

used like simple, except it restricts the type-level lists involved to have the same length, without fixing the type of container or the elements in the list.

asLengthOf :: SameLength x y => r x -> s y -> r x

type family SameLengths xs :: Constraint

Labels

class SameLabels x y

sameLabels :: SameLabels x y => p (r x) (f (q y)) -> p (r x) (f (q y))

sameLabels constrains the type of an optic, such that the labels (t in Tagged t a) are the same. x or y may have more elements than the other, in which case the elements at the end of the longer list do not have their labels constrained.

see also sameLength

A list has only Tagged values

class HAllTaggedLV ps

The Record, Variant, TIP, TIC type constructors only make sense when they are applied to an instance of this class

type family ZipTagged ts vs :: [*]

see Data.HList.Record.zipTagged

module Data.Proxy

module Data.Tagged

class Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

mappend mempty x = x

mappend x mempty = x

mappend x (mappend y z) = mappend (mappend x y) z

mconcat = foldr mappend mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Minimal complete definition: mempty and mappend.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

Minimal complete definition

mempty, mappend

Methods

mempty :: a

Identity of mappend

mappend :: a -> a -> a

An associative operation

mconcat :: [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.