Data.HList.HList

Description

The HList library

(C) 2004, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Basic declarations for typeful heterogeneous lists.

Excuse the unstructured haddocks: while there are many declarations here some are alternative implementations should be grouped, and the definitions here are analgous to many list functions in the Prelude.

Heterogeneous type sequences

There are three sensible ways to define HLists:

data HList (l::[*]) where
    HNil  :: HList '[]
    HCons :: e -> HList l -> HList (e ': l)

This ensures that sequences can only be formed with Nil and Cons. The argument to HList is a promoted lists (kind [*]), which has a more attractive syntax.

Earlier versions of HList used an algebraic data type:

data HCons a b = HCons a b
data HNil = HNil

Disadvantages:

values with types like HCons Int Double to be created, which are nonsense to the functions in HList

some recursive functions do not need a class with the GADT. For example:

   hInit :: HListGADT (x ': xs) -> HListGADT (HInit (x ': xs))
   hInit (HCons x xs@(HCons _ _)) = HCons x (hInit xs)
   hInit (HCons _ HNil) = HNil

   type family HInit (xs :: [k]) :: [k]
 

but without the GADT, hInit is written as in a class, which complicates inferred types

Advantages

lazy pattern matches are allowed, so lazy pattern matching on a value undefined :: HList [a,b,c] can create the spine of the list. hProxies avoids the use of undefined, but a slightly more complicated class context has to be written or inferred.

type inference is better if you want to directly pattern match see stackoverflow post here

better pattern exhaustiveness checking (as of ghc-7.8)

standalone deriving works

Data.Coerce.coerce works because the parameters have role representational, not nominal as they are for the GADT and data family. Probably the GADT/type family actually do have a representational role: http://stackoverflow.com/questions/24222552/does-this-gadt-actually-have-type-role-representational

The data family version (currently used) gives the same type constructor HList :: [*] -> * as the GADT, while pattern matching behaves like the algebraic data type. Furthermore, nonsense values like HCons 1 2 :: HCons Int Int cannot be written with the data family.

A variation on the data family version is

data instance HList '[] = HNil
newtype instance HList (x ': xs) = HCons1 (x, HList xs)
pattern HCons x xs = HCons1 (x, xs)

This allows HList to have a nominal role, but on the other hand the PatternSynonym is not supported with ghc-7.6 and exhaustiveness checking is not as good (warnings for _ being unmatched)

data family HList l

class HProxiesFD xs pxs | pxs -> xs, xs -> pxs where

creates a HList of Proxies

Methods

hProxies :: HList pxs

type HProxies xs = HProxiesFD xs (AddProxy xs)

type family AddProxy xs :: k

Add Proxy to a type

>>> let x = undefined :: HList (AddProxy [Char,Int])
>>> :t x
x :: HList '[Proxy Char, Proxy Int]

type family DropProxy xs :: k

inverse of AddProxy

data ReadElement

Constructors

ReadElement

Basic list functions

hHead :: HList (e : l) -> e

head

hTail :: HList (e : l) -> HList l

tail

hLast :: HRevApp l1 ([] *) ((:) * e l) => HList l1 -> e

last

class HInit xs where

Associated Types

type HInitR xs :: [*]

Methods

hInit :: HList xs -> HList (HInitR xs)

type family HLength x :: HNat

Length, but see HLengthEq instead

hLength :: HLengthEq l n => HList l -> Proxy n

Append

type family HAppendListR l1 l2 :: [k]

class HAppendList l1 l2 where

Methods

hAppendList :: HList l1 -> HList l2 -> HList (HAppendListR l1 l2)

the same as hAppend

Alternative append

append' :: [a] -> [a] -> [a]

hAppend' below is implemented using the same idea

hAppend' :: HFoldr FHCons v l r => HList l -> v -> r

Alternative implementation of hAppend. Demonstrates HFoldr

data FHCons

Constructors

FHCons

Historical append

The original HList code is included below. In both cases we had to program the algorithm twice, at the term and the type levels.

The class HAppend

class HAppend l l' l'' | l l' -> l''
 where
  hAppend :: l -> l' -> l''

The instance following the normal append

instance HList l => HAppend HNil l l
 where
  hAppend HNil l = l

instance (HList l, HAppend l l' l'')
      => HAppend (HCons x l) l' (HCons x l'')
 where
  hAppend (HCons x l) l' = HCons x (hAppend l l')

Reversing HLists

type family HRevAppR l1 l2 :: [k]

class HRevApp l1 l2 l3 | l1 l2 -> l3 where

Methods

hRevApp :: HList l1 -> HList l2 -> HList l3

class HReverse xs sx | xs -> sx, sx -> xs where

Methods

hReverse :: HList xs -> HList sx

hReverse_ :: HRevApp l1 ([] *) l3 => HList l1 -> HList l3

a version of hReverse that does not allow the type information to flow backwards

A nicer notation for lists

hEnd :: HList l -> HList l

Note:

x :: HList a

means: forall a. x :: HList a

hEnd x

means: exists a. x :: HList a

List termination

hBuild :: HBuild' [] r => r

Building lists

class HBuild' l r where

Methods

hBuild' :: HList l -> r

examples

The classes above allow the third (shortest) way to make a list (containing a,b,c) in this case

list = a `HCons` b `HCons` c `HCons` HNil
list = a .*. b .*. c .*. HNil
list = hEnd $ hBuild a b c

>>> let x = hBuild True in hEnd x
H[True]

>>> let x = hBuild True 'a' in hEnd x
H[True,'a']

>>> let x = hBuild True 'a' "ok" in hEnd x
H[True,'a',"ok"]

hBuild can also produce a Record, such that

hBuild x y ^. from unlabeled

can also be produced using

hEndR $ hBuild x y

historical

the show instance has since changed, but these uses of 'hBuild'/'hEnd' still work

HList> let x = hBuild True in hEnd x
HCons True HNil

HList> let x = hBuild True 'a' in hEnd x
HCons True (HCons 'a' HNil)

HList> let x = hBuild True 'a' "ok" in hEnd x
HCons True (HCons 'a' (HCons "ok" HNil))

HList> hEnd (hBuild (Key 42) (Name "Angus") Cow (Price 75.5))
HCons (Key 42) (HCons (Name "Angus") (HCons Cow (HCons (Price 75.5) HNil)))

HList> hEnd (hBuild (Key 42) (Name "Angus") Cow (Price 75.5)) == angus
True

folds

foldr

Consume a heterogenous list.

class HFoldr f v l r where

Methods

hFoldr :: f -> v -> HList l -> r

class HScanr f z ls rs where

Methods

hScanr :: f -> z -> HList ls -> HList rs

class HFoldr1 f l r where

Methods

hFoldr1 :: f -> HList l -> r

foldl

class HFoldl f z xs r where

like foldl

>>> hFoldl (uncurry $ flip (:)) [] (1 `HCons` 2 `HCons` HNil)
[2,1]

Methods

hFoldl :: f -> z -> HList xs -> r

unfolds

unfold

Produce a heterogenous list. Uses the more limited Apply instead of App since that's all that is needed for uses of this function downstream. Those could in principle be re-written.

hUnfold :: (HUnfoldFD f (ApplyR f a) z, Apply f a) => f -> a -> HList z

type HUnfold p s = HUnfoldR p (ApplyR p s)

type family HUnfoldR p res :: [*]

type HUnfold' p res = HUnfoldFD p (ApplyR p res) (HUnfold p res)

class HUnfoldFD p res z | p res -> z where

Methods

hUnfold' :: p -> res -> HList z

replicate

class HLengthEq es n => HReplicateFD n e es | n e -> es, es -> n where

Sometimes the result type can fix the type of the first argument:

>>> hReplicate Proxy () :: HList '[ (), (), () ]
H[(),(),()]

However, with HReplicate all elements must have the same type, so it may be easier to use HList2List:

>>> list2HList (repeat 3) :: Maybe (HList [Int, Int, Int])
Just H[3,3,3]

Methods

hReplicate :: Proxy n -> e -> HList es

type HReplicate n e = HReplicateFD n e (HReplicateR n e)

type family HReplicateR n e :: [k]

would be associated with HReplicate except we want it to work with e of any kind, not just * that you can put into a HList. An "inverse" of HLength

class HLengthEq r n => HReplicateF n f z r | r -> n where

HReplicate produces lists that can be converted to ordinary lists

>>> let two = hSucc (hSucc hZero)
>>> let f = Fun' fromInteger :: Fun' Num Integer

>>> :t applyAB f
applyAB f :: Num b => Integer -> b

>>> hReplicateF two f 3
H[3,3]

>>> hReplicateF Proxy f 3 :: HList [Int, Double, Integer]
H[3,3.0,3]

Methods

hReplicateF :: HLengthEq r n => Proxy n -> f -> z -> HList r

iterate

class HLengthEq r n => HIterate n f z r where

This function behaves like iterate, with an extra argument to help figure out the result length

>>> let three = hSucc (hSucc (hSucc hZero))
>>> let f = Fun Just :: Fun '() Maybe

>>> :t applyAB f
applyAB f :: a -> Maybe a

f is applied to different types:

>>> hIterate three f ()
H[(),Just (),Just (Just ())]

It is also possible to specify the length later on, as done with Prelude.iterate

>>> let take3 x | _ <- hLength x `asTypeOf` three = x
>>> take3 $ hIterate Proxy f ()
H[(),Just (),Just (Just ())]

Methods

hIterate :: HLengthEq r n => Proxy n -> f -> z -> HList r

concat

type HConcat xs = HConcatFD xs (HConcatR xs)

Like concat but for HLists of HLists.

Works in ghci... puzzling as what is different in doctest (it isn't -XExtendedDefaultRules)

>>> let a = hEnd $ hBuild 1 2 3
>>> let b = hEnd $ hBuild 'a' "abc"
>>> hConcat $ hBuild a b
H[1, 2, 3, 'a', "abc"]

hConcat :: HConcat xs => HList xs -> HList (HConcatR xs)

type family HConcatR a :: [*]

type family UnHList a :: [*]

class HConcatFD xxs xs | xxs -> xs where

Methods

hConcatFD :: HList xxs -> HList xs

class HAppendFD a b ab | a b -> ab where

Methods

hAppendFD :: HList a -> HList b -> HList ab

traversing HLists

producing HList

map

It could be implemented with hFoldr, as we show further below

newtype HMap f

hMap is written such that the length of the result list can be determined from the length of the argument list (and the other way around). Similarly, the type of the elements of the list is propagated in both directions too.

>>> :set -XNoMonomorphismRestriction
>>> let xs = 1 .*. 'c' .*. HNil
>>> :t hMap (HJust ()) xs
hMap (HJust ()) xs :: Num y => HList '[HJust y, HJust Char]

These 4 examples show that the constraint on the length (2 in this case) can be applied before or after the hMap. That inference is independent of the direction that type information is propagated for the individual elements.

>>> let asLen2 xs = xs `asTypeOf` (undefined :: HList '[a,b])

>>> let lr xs = asLen2 (applyAB (HMap HRead) xs)
>>> let ls xs = asLen2 (applyAB (HMap HShow) xs)
>>> let rl xs = applyAB (HMap HRead) (asLen2 xs)
>>> let sl xs = applyAB (HMap HShow) (asLen2 xs)

>>> :t lr
lr
  :: (Read ..., Read ...) => HList '[String, String] -> HList '[..., ...]

>>> :t rl
rl
  :: (Read ..., Read ...) => HList '[String, String] -> HList '[..., ...]

>>> :t ls
ls
  :: (Show ..., Show ...) => HList '[..., ...] -> HList '[String, String]

>>> :t sl
sl
  :: (Show ..., Show ...) => HList '[..., ...] -> HList '[String, String]

Constructors

HMap f

hMap :: (HMapAux r f a b, SameLength' * * b a, SameLength' * * a b) => f -> r a -> r b

hMapL :: (HMapAux HList f a b, SameLength' * * b a, SameLength' * * a b) => f -> HList a -> HList b

hMap constrained to HList

newtype HMapL f

Constructors

HMapL f

class (SameLength a b, HMapAux r f a b) => HMapCxt r f a b

class HMapAux r f x y where

Methods

hMapAux :: SameLength x y => f -> r x -> r y

alternative implementation

currently broken

newtype MapCar f

Constructors

MapCar f

hMapMapCar :: HFoldr (MapCar f) (HList []) l l' => f -> HList l -> l'

Same as hMap only a different implementation.

appEndo . mconcat . map Endo

hComposeList :: HFoldr Comp (a -> a) l (t -> a) => HList l -> t -> a

>>> let xs = length .*. (+1) .*. (*2) .*. HNil
>>> hComposeList xs "abc"
8

sequence

class (Applicative m, SameLength a b) => HSequence m a b | a -> b, m b -> a where

A heterogeneous version of

sequenceA :: (Applicative m) => [m a] -> m [a]

Only now we operate on heterogeneous lists, where different elements may have different types a. In the argument list of monadic values (m a_i), although a_i may differ, the monad m must be the same for all elements. That's why we needed Data.HList.TypeCastGeneric2 (currently (~)). The typechecker will complain if we attempt to use hSequence on a HList of monadic values with different monads.

The hSequence problem was posed by Matthias Fischmann in his message on the Haskell-Cafe list on Oct 8, 2006

http://www.haskell.org/pipermail/haskell-cafe/2006-October/018708.html

http://www.haskell.org/pipermail/haskell-cafe/2006-October/018784.html

Maybe

>>> hSequence $ Just (1 :: Integer) `HCons` (Just 'c') `HCons` HNil
Just H[1,'c']

>>> hSequence $  return 1 `HCons` Just  'c' `HCons` HNil
Just H[1,'c']

List

>>> hSequence $ [1] `HCons` ['c'] `HCons` HNil
[H[1,'c']]

Methods

hSequence :: HList a -> m (HList b)

alternative implementation

hSequence2 :: (HFoldr (LiftA2 FHCons) (f (HList ([] *))) l (f a), Applicative f) => HList l -> f a

hSequence2 is not recommended over hSequence since it possibly doesn't allow inferring argument types from the result types. Otherwise this version should do exactly the same thing.

The DataKinds version needs a little help to find the type of the return HNil, unlike the original version, which worked just fine as

hSequence l = hFoldr ConsM (return HNil) l

producing homogenous lists

map (no sequencing)

This one we implement via hFoldr

newtype Mapcar f

Constructors

Mapcar f

type HMapOut f l e = HFoldr (Mapcar f) [e] l [e]

hMapOut :: forall f e l. HMapOut f l e => f -> HList l -> [e]

compare hMapOut f with hList2List . hMap f

mapM

hMapM :: (Monad m, HMapOut f l (m e)) => f -> HList l -> [m e]

mapM :: forall b m a. (Monad m) => (a -> m b) -> [a] -> m [b]

Likewise for mapM_.

See hSequence if the result list should also be heterogenous.

hMapM_ :: (Monad m, HMapOut f l (m ())) => f -> HList l -> m ()

GHC doesn't like its own type.

hMapM_ :: forall m a f e. (Monad m, HMapOut f a (m e)) => f -> a -> m ()

Without explicit type signature, it's Ok. Sigh. Anyway, Hugs does insist on a better type. So we restrict as follows:

Ensure a list to contain HNats only

type family HNats l :: [HNat]

We do so constructively, converting the HList whose elements are Proxy HNat to [HNat]. The latter kind is unpopulated and is present only at the type level.

hNats :: HList l -> Proxy (HNats l)

Membership tests

class HMember e1 l b | e1 l -> b

Check to see if an HList contains an element with a given type This is a type-level only test

class HMember' b0 e1 l b | b0 e1 l -> b

type family HMemberP pred e1 l :: Bool

The following is a similar type-only membership test It uses the user-supplied curried type equality predicate pred

type family HMemberP' pred e1 l pb :: Bool

hMember :: HMember e l b => Proxy e -> Proxy l -> Proxy b

Another type-level membership test

class HMemberM e1 l r | e1 l -> r

Check to see if an element e occurs in a list l If not, return 'Nothing If the element does occur, return 'Just l1 where l1 is a type-level list without e

class HMemberM1 b e1 l r | b e1 l -> r

class HMemberM2 b e1 l r | b e1 l -> r

Staged equality for lists

removed. use Typeable instead

Find an element in a set based on HEq

class HFind1 e l n => HFind e l n | e l -> n

It is a pure type-level operation

class HFind1 e l n | e l -> n

class HFind2 b e l n | b e l -> n

Membership test based on type equality

class HTMember e l b | e l -> b

could be an associated type if HEq had one

hTMember :: HTMember e l b => e -> HList l -> Proxy b

Intersection based on HTMember

class HTIntersect l1 l2 l3 | l1 l2 -> l3 where

Methods

hTIntersect :: HList l1 -> HList l2 -> HList l3

Like intersect

class HTIntersectBool b h t l1 l2 | b h t l1 -> l2 where

Methods

hTIntersectBool :: Proxy b -> h -> HList t -> HList l1 -> HList l2

Convert between heterogeneous lists and homogeneous ones

class HList2List l e | l -> e where

hMapOut id is similar, except this function is restricted to HLists that actually contain a value (so the list produced will be nonempty). This restriction allows adding a functional dependency, which means that less type annotations can be necessary.

Methods

hList2List :: HList l -> [e]

list2HListSuffix :: [e] -> Maybe (HList l, [e])

list2HList :: HList2List l e => [e] -> Maybe (HList l)

listAsHList :: (HList2List l1 e1, HList2List l e, Choice p, Applicative f) => p (HList l1) (f (HList l)) -> p [e1] (f [e])

Prism [s] [t] (HList s) (HList t)

listAsHList' :: (HList2List l e, Choice p, Applicative f) => p (HList l) (f (HList l)) -> p [e] (f [e])

Prism' [a] (HList s)

where s ~ HReplicateR n a

With HMaybe

Turn list in a list of justs

class (FromHJustR (ToHJustR l) ~ l) => ToHJust l where

the same as map Just

>>> toHJust (2 .*. 'a' .*. HNil)
H[HJust 2,HJust 'a']

>>> toHJust2 (2 .*. 'a' .*. HNil)
H[HJust 2,HJust 'a']

Associated Types

type ToHJustR l :: [*]

Methods

toHJust :: HList l -> HList (ToHJustR l)

toHJust2 :: (HMapCxt r (HJust ()) a b, ToHJust a, b ~ ToHJustR a) => r a -> r b

alternative implementation. The Apply instance is in Data.HList.FakePrelude. A longer type could be inferred.

Extract justs from list of maybes

class (FromHJustR (ToHJustR l) ~ l) => FromHJust l where

Associated Types

type FromHJustR l :: [*]

Methods

fromHJust :: HList l -> HList (FromHJustR l)

alternative implementation

fromHJust2 :: HMapCxt r HFromJust a b => r a -> r b

This implementation is shorter.

data HFromJust

Constructors

HFromJust

Annotated lists

data HAddTag t

Constructors

HAddTag t

data HRmTag

Constructors

HRmTag

hAddTag :: (HMapAux r (HAddTag t) a b, SameLength' * * b a, SameLength' * * a b) => t -> r a -> r b

hRmTag :: (HMapAux r HRmTag a b, SameLength' * * b a, SameLength' * * a b) => r a -> r b

hFlag :: (HMapAux r (HAddTag (Proxy Bool True)) a b, SameLength' * * b a, SameLength' * * a b) => r a -> r b

Annotate list with a type-level Boolean

hFlag :: HMapCxt (HAddTag (Proxy True)) l r => HList l -> HList r

Splitting by HTrue and HFalse

class HSplit l where

Analogus to Data.List.partition snd. See also HPartition

>>> let (.=.) :: p x -> y -> Tagged x y; _ .=. y = Tagged y
>>> hSplit $ hTrue .=. 2 .*. hTrue .=. 3 .*. hFalse .=. 1 .*. HNil
(H[2,3],H[1])

it might make more sense to instead have LVPair Bool e instead of (e, Proxy Bool) since the former has the same runtime representation as e

Associated Types

type HSplitT l :: [*]

type HSplitF l :: [*]

Methods

hSplit :: HList l -> (HList (HSplitT l), HList (HSplitF l))

Splitting by Length

class (HLengthEq xs n, HAppendList1 xs ys xsys) => HSplitAt n xsys xs ys | n xsys -> xs ys, xs ys -> xsys, xs -> n where

splitAt

setup

>>> let two = hSucc (hSucc hZero)
>>> let xsys = hEnd $ hBuild 1 2 3 4

If a length is explicitly provided, the resulting lists are inferred

>>> hSplitAt two xsys
(H[1,2],H[3,4])

>>> let sameLength_ :: SameLength a b => r a -> r b -> r a; sameLength_ = const
>>> let len2 x = x `sameLength_` HCons () (HCons () HNil)

If the first chunk of the list (a) has to be a certain length, the type of the Proxy argument can be inferred.

>>> case hSplitAt Proxy xsys of (a,b) -> (len2 a, b)
(H[1,2],H[3,4])

Methods

hSplitAt :: Proxy n -> HList xsys -> (HList xs, HList ys)

class HSplitAt1 accum n xsys xs ys | accum n xsys -> xs ys where

helper for HSplitAt

Methods

hSplitAt1 :: HList accum -> Proxy n -> HList xsys -> (HList xs, HList ys)

class (SameLength' (HReplicateR n ()) xs, HLengthEq1 xs n, HLengthEq2 xs n) => HLengthEq xs n | xs -> n

a better way to write HLength xs ~ n because:

it works properly with ghc-7.10 (probably another example of ghc bug #10009)

it works backwards a bit in that if n is known, then xs can be refined:

>>> undefined :: HLengthEq xs HZero => HList xs
H[]

class HLengthEq1 xs n

class HLengthEq2 xs n | xs -> n

class HStripPrefix xs xsys ys => HAppendList1 xs ys xsys | xs ys -> xsys, xs xsys -> ys

HAppendList1 xs ys xsys is the type-level way of saying xs ++ ys == xsys

used by HSplitAt

class HStripPrefix xs xsys ys | xs xsys -> ys

analog of stripPrefix

Conversion to and from tuples

class HTuple v t | v -> t, t -> v where

Methods

hToTuple :: HList v -> t

alternatively: hUncurry (,,,)

hFromTuple :: t -> HList v

hTuple :: (HTuple v1 b, HTuple v a, Profunctor p, Functor f) => p a (f b) -> p (HList v) (f (HList v1))

Iso (HList v) (HList v') a b

hTuple' :: (HTuple v b, Profunctor p, Functor f) => p b (f b) -> p (HList v) (f (HList v))

Iso' (HList v) a

class HTails a b | a -> b, b -> a where

tails

Methods

hTails :: HList a -> HList b

class HInits a b | a -> b, b -> a where

inits

Methods

hInits :: HList a -> HList b

class HInits1 a b | a -> b, b -> a where

behaves like tail . inits

Methods

hInits1 :: HList a -> HList b

data FHCons2 x

similar to FHCons

Constructors

FHCons2 x

type family HMapCons x xxs :: [*]

evidence to satisfy the fundeps in HInits

type family HMapTail xxs :: [*]

evidence to satisfy the fundeps in HInits

partition

class HPartitionEq f x1 xs xi xo | f x1 xs -> xi xo where

HPartitionEq f x1 xs xi xo is analogous to

(xi,xo) = partition (f x1) xs

where f is a "function" passed in using it's instance of HEqBy

Methods

hPartitionEq :: Proxy f -> Proxy x1 -> HList xs -> (HList xi, HList xo)

class HPartitionEq1 b f x1 x xs xi xo | b f x1 x xs -> xi xo where

Methods

hPartitionEq1 :: Proxy b -> Proxy f -> Proxy x1 -> x -> HList xs -> (HList xi, HList xo)

groupBy

class HGroupBy f as gs | f as -> gs, gs -> as where

HGroupBy f x y is analogous to y = groupBy f x

given that f is used by HEqBy

Methods

hGroupBy :: Proxy f -> HList as -> HList gs

span

class HSpanEqBy f x y fst snd | f x y -> fst snd, fst snd -> y where

HSpanEq x y fst snd is analogous to (fst,snd) = span (== x) y

Methods

hSpanEqBy :: Proxy f -> x -> HList y -> (HList fst, HList snd)

class HSpanEqBy1 f x y i o | f x y -> i o where

Methods

hSpanEqBy1 :: Proxy f -> x -> HList y -> (HList i, HList o)

class HSpanEqBy2 b f x y ys i o | b f x y ys -> i o where

Methods

hSpanEqBy2 :: Proxy b -> Proxy f -> x -> y -> HList ys -> (HList i, HList o)

zip

see alternative implementations in Data.HList.HZip

class HZipList x y l | x y -> l, l -> x y where

Methods

hZipList :: HList x -> HList y -> HList l

hUnzipList :: HList l -> (HList x, HList y)

Monoid instance

helper functions

data ConstMempty

Constructors

ConstMempty

data UncurryMappend

Constructors

UncurryMappend