Data.HList.Record

Description

The HList library

(C) 2004-2006, Oleg Kiselyov, Ralf Laemmel, Keean Schupke

Extensible records

The three-ish models of labels that go with this module;

Data.HList.Label3

Data.HList.Label5

Data.HList.Label6

Data.HList.Labelable

These used to work:

Data.HList.Label1

Data.HList.Label2

Data.HList.Label4

labels used for doctests

>>> let x = Label :: Label "x"
>>> let y = Label :: Label "y"
>>> let z = Label :: Label "z"

Records

Labels

Record types as label-value pairs, where label is purely phantom. Thus the run-time representation of a field is the same as that of its value, and the record, at run-time, is indistinguishable from the HList of field values. At run-time, all information about the labels is erased.

The type from Data.Tagged is used.

module Data.Tagged

(.=.) :: Label l -> v -> Tagged l v infixr 4

Create a value with the given label. Analagous to a data constructor such as Just, Left, or Right. Higher fixity than record-modification operations like (.*.), (.-.), etc. to support expression like the below w/o parentheses:

>>> x .=. "v1" .*. y .=. '2' .*. emptyRecord
Record{x="v1",y='2'}

Record

newtype Record r

Constructors

Record (HList r)

mkRecord :: HRLabelSet r => HList r -> Record r

Build a record

emptyRecord :: Record []

Build an empty record

hEndR :: Record a -> Record a

serves the same purpose as hEnd

hEndP :: Proxy (xs :: [k]) -> Proxy xs

hEndP $ hBuild label1 label2

is one way to make a Proxy of labels (for use with asLabelsOf for example). Another way is

label1 .*. label2 .*. emptyProxy

hListRecord :: (HLabelSet [*] (LabelsOf r), HAllTaggedLV r, Profunctor p, Functor f) => p (Record r) (f (Record t)) -> p (HList r) (f (HList t))

HRLabelSet t => Iso (HList s) (HList t) (Record s) (Record t)

hListRecord' :: (Profunctor p, Functor f) => p (Record r) (f (Record r)) -> p (HList r) (f (HList r))

Iso' (HList s) (Record s)

Getting Labels

type family LabelsOf ls :: [*]

Construct the (phantom) list of labels of a record, or list of Label.

labelsOf :: hlistOrRecord l -> Proxy (LabelsOf l)

asLabelsOf :: (HAllTaggedLV x, SameLabels x y, SameLength x y) => r x -> s y -> r x

similar to asTypeOf:

>>> let s0 = Proxy :: Proxy '["x", "y"]
>>> let s1 = Proxy :: Proxy '[Label "x", Label "y"]
>>> let s2 = Proxy :: Proxy '[Tagged "x" Int, Tagged "y" Char]

>>> let f0 r = () where _ = r `asLabelsOf` s0
>>> let f1 r = () where _ = r `asLabelsOf` s1
>>> let f2 r = () where _ = r `asLabelsOf` s2

>>> :t f0
f0 :: r '[Tagged "x" v, Tagged "y" v1] -> ()

>>> :t f1
f1 :: r '[Tagged "x" v, Tagged "y" v1] -> ()

>>> :t f2
f2 :: r '[Tagged "x" v, Tagged "y" v1] -> ()

Getting Values

class SameLength r (RecordValuesR r) => RecordValues r where

Construct the HList of values of the record.

Associated Types

type RecordValuesR r :: [*]

Methods

recordValues' :: HList r -> HList (RecordValuesR r)

recordValues :: RecordValues r => Record r -> HList (RecordValuesR r)

hMapTaggedFn :: HMapTaggedFn a b => HList a -> Record b

"inverse" to recordValues

unlabeled0 :: (RecordValues y, RecordValues x, HMapAux HList TaggedFn (RecordValuesR y) y, SameLabels [*] [*] x y, Profunctor p, Functor f) => p (HList (RecordValuesR x)) (f (HList (RecordValuesR y))) -> p (Record x) (f (Record y))

Iso (Record s) (Record t) (HList a) (HList b)

view unlabeled == recordValues

type Unlabeled x y = (HMapCxt HList TaggedFn (RecordValuesR y) y, RecordValues x, RecordValues y, SameLength (RecordValuesR x) (RecordValuesR y), SameLength x y, SameLabels x y, HAllTaggedLV x, HAllTaggedLV y)

unlabeled :: (Unlabeled x y, Profunctor p, Functor f) => (HList (RecordValuesR x) `p` f (HList (RecordValuesR y))) -> Record x `p` f (Record y)

type Unlabeled' x = Unlabeled x x

unlabeled' :: (Unlabeled' x, Profunctor p, Functor f) => (HList (RecordValuesR x) `p` f (HList (RecordValuesR x))) -> Record x `p` f (Record x)

Unlabeled' x => Iso' (Record x) (HList (RecordValuesR x))

Operations

Show

A corresponding Show instance exists as

show x = "Record {" ++ showComponents "" x ++ "}"

class ShowComponents l where

Methods

showComponents :: String -> HList l -> String

class ShowLabel l where

Methods

showLabel :: Label l -> String

Extension

hExtend, hAppend

(.*.) :: HExtend e l => e -> l -> HExtendR e l

Delete

hDeleteAtLabel label record

(.-.) :: HDeleteAtLabel r l xs xs' => r xs -> Label l -> r xs' infixl 2

Remove a field from a record. At the same level as other record modification options (.*.). Analagous to (\\) in lists.

record1 .-. label1

label1 .=. value1 .*.
label2 .=. value2 .-.
label2 .*.
emptyRecord

label1 .=. value1 .-.
label1 .*.
label2 .=. value2 .*.
emptyRecord

record1 .*. label1 .=. record2 .!. label1
        .*. label2 .=. record2 .!. label2
        .-. label1

class HDeleteLabels ks r r' | ks r -> r' where

Methods

hDeleteLabels

Arguments

:: proxy (ks :: [*])
as provided by labelsOf

-> Record r

-> Record r'

Lookup/update

Lens-based setters/getters are popular. hLens packages up hUpdateAtLabel and hLookupByLabel.

Refer to examples/lens.hs and examples/labelable.hs for examples.

class HLensCxt x r s t a b => HLens x r s t a b | x s b -> t, x t a -> s, x s -> a, x t -> b where

Methods

hLens :: Label x -> forall f. Functor f => (a -> f b) -> r s -> f (r t)

hLens :: Label x -> Lens (r s) (r t) a b

Lookup

class HasField l r v | l r -> v where

This is a baseline implementation. We use a helper class, HasField, to abstract from the implementation.

Because hLookupByLabel is so frequent and important, we implement it separately, more efficiently. The algorithm is familiar assq, only the comparison operation is done at compile-time

Methods

hLookupByLabel :: Label l -> r -> v

class HasFieldM l r v | l r -> v where

a version of HasField hLookupByLabel .!. that returns a default value when the label is not in the record:

>>> let r = x .=. "the x value" .*. emptyRecord

>>> hLookupByLabelM y r ()
()

>>> hLookupByLabelM x r ()
"the x value"

Methods

hLookupByLabelM

Arguments

:: Label l

-> r
Record (or Variant,TIP,TIC)

-> t
default value

-> DemoteMaybe t v

(.!.) :: HasField l r v => r -> Label l -> v infixr 9

Lookup a value in a record by its label. Analagous to (!!), the list indexing operation. Highest fixity, like (!!).

>>> :{
let record1 = x .=. 3 .*.
              y .=. 'y' .*.
              emptyRecord
:}

>>> record1 .!. x
3

>>> record1 .!. y
'y'

>>> :{
let r2 = y .=. record1 .!. x .*.
         z .=. record1 .!. y .*.
         emptyRecord
:}

>>> r2
Record{y=3,z='y'}

Note that labels made following Data.HList.Labelable allow using "Control.Lens.^." instead.

Update

(.@.) :: (HUpdateAtLabel k record l v r r', SameLength' * * r' r, SameLength' * * r r') => Tagged k l v -> record r -> record r' infixr 2

Update a field with a particular value. Same fixity as (.*.) so that extensions and updates can be chained. There is no real list analogue, since there is no Prelude defined update.

label1 .=. value1 .@. record1

class HUpdateAtLabel record l v r r' | l v r -> r', l r' -> v where

hUpdateAtLabel label value record

Methods

hUpdateAtLabel :: SameLength r r' => Label l -> v -> record r -> record r'

type-preserving versions

Note: these restrict the resulting record type to be the same as in input record type, which can help reduce the number of type annotations needed

(.<.) :: (HUpdateAtLabel k record l v r r, SameLength' * * r r) => Tagged k l v -> record r -> record r infixr 2

The same as .@., except type preserving. It has the same fixity as (.@.).

type HTPupdateAtLabel record l v r = (HUpdateAtLabel record l v r r, SameLength' r r)

hTPupdateAtLabel :: HTPupdateAtLabel record l v r => Label l -> v -> record r -> record r

We could also say:

hTPupdateAtLabel l v r = hUpdateAtLabel l v r `asTypeOf` r

Then we were taking a dependency on Haskell's type equivalence. This would also constrain the actual implementation of hUpdateAtLabel.

A variation on hUpdateAtLabel: type-preserving update.

Rename Label

hRenameLabel :: (HasField k l (r v) v1, HExtend (Tagged k1 l1 v1) (r v'), HDeleteAtLabel k r l v v') => Label k l -> Label k1 l1 -> r v -> HExtendR (Tagged k1 l1 v1) (r v')

Rename the label of record

>>> hRenameLabel x y (x .=. () .*. emptyRecord)
Record{y=()}

Projection

type family Labels xs :: *

A helper to make the Proxy needed by hProjectByLabels, and similar functions which accept a list of kind [*].

For example:

(rin,rout) = hProjectByLabels2 (Proxy :: Labels ["x","y"]) r

behaves like

rin = r .!. (Label :: Label "x") .*.
      r .!. (Label :: Label "y") .*.
      emptyRecord

rout = r .-. (Label :: Label "x") .-. (Label :: Label "y")

It is also an important operation: the basis of many deconstructors -- so we try to implement it efficiently.

hProjectByLabels :: (HRLabelSet a, H2ProjectByLabels ls t a b) => proxy ls -> Record t -> Record a

hProjectByLabels ls r returns r with only the labels in ls remaining

hProjectByLabels' :: (HLabelSet [*] (LabelsOf l), HLabelSet [*] (LabelsOf r), H2ProjectByLabels (LabelsOf l) t r b, HRearrange3 (LabelsOf l) r l, SameLength' * * (LabelsOf l) r, SameLength' * * l r, SameLength' * * r (LabelsOf l), SameLength' * * r l, HAllTaggedLV r) => Record t -> Record l

hProjectByLabels2 :: (H2ProjectByLabels ls t t1 t2, HRLabelSet t1, HRLabelSet t2) => Proxy ls -> Record t -> (Record t1, Record t2)

See H2ProjectByLabels

a lens for projection

see Data.HList.Labelable.Projected

Unions

Left

class HLeftUnion r r' r'' | r r' -> r'' where

Methods

hLeftUnion :: Record r -> Record r' -> Record r''

(.<++.) :: HLeftUnion r r' r'' => Record r -> Record r' -> Record r'' infixl 1

Similar to list append, so give this slightly lower fixity than (.*.), so we can write:

field1 .=. value .*. record1 .<++. record2

Symmetric

Compute the symmetric union of two records r1 and r2 and return the pair of records injected into the union (ru1, ru2).

To be more precise, we compute the symmetric union type ru of two record types r1 and r2. The emphasis on types is important.

The two records (ru1,ru2) in the result of unionSR have the same type ru, but they are generally different values. Here the simple example: suppose

 r1 = (Label .=. True)  .*. emptyRecord
 r2 = (Label .=. False) .*. emptyRecord

Then unionSR r1 r2 will return (r1,r2). Both components of the result are different records of the same type.

To project from the union ru, use hProjectByLabels. It is possible to project from the union obtaining a record that was not used at all when creating the union.

We do assure however that if unionSR r1 r2 gave (r1u,r2u), then projecting r1u onto the type of r1 gives the value identical to r1. Ditto for r2.

class UnionSymRec r1 r2 ru | r1 r2 -> ru where

Methods

unionSR :: Record r1 -> Record r2 -> (Record ru, Record ru)

Reorder Labels

hRearrange :: (HLabelSet ls, HRearrange ls r r') => Proxy ls -> Record r -> Record r'

Rearranges a record by labels. Returns the record r, rearranged such that the labels are in the order given by ls. (LabelsOf r) must be a permutation of ls.

hRearrange' :: (HLabelSet [*] (LabelsOf l), HRearrange3 (LabelsOf l) r l, SameLength' * * (LabelsOf l) r, SameLength' * * l r, SameLength' * * r (LabelsOf l), SameLength' * * r l) => Record r -> Record l

hRearrange' is hRearrange where ordering specified by the Proxy argument is determined by the result type.

With built-in haskell records, these e1 and e2 have the same type:

data R = R { x, y :: Int }
e1 = R{ x = 1, y = 2}
e2 = R{ y = 2, x = 1}

hRearrange' can be used to allow either ordering to be accepted:

h1, h2 :: Record [ Tagged "x" Int, Tagged "y" Int ]
h1 = hRearrange' $
    x .=. 1 .*.
    y .=. 2 .*.
    emptyRecord

h2 = hRearrange' $
    y .=. 2 .*.
    x .=. 1 .*.
    emptyRecord

isos using hRearrange

class Rearranged r s t a b where

Methods

rearranged :: (Profunctor p, Functor f) => (r a `p` f (r b)) -> r s `p` f (r t)

rearranged' :: (Rearranged k r s s b b, Profunctor p, Functor f) => p (r b) (f (r b)) -> p (r s) (f (r s))

Iso' (r s) (r a)

where s is a permutation of a

Apply a function to all values

hMapR :: (HMapAux HList (HFmap f) x y, SameLength' * * y x, SameLength' * * x y) => f -> Record x -> Record y

map over the values of a record. This is a shortcut for

\ f (Record a) -> Record (hMap (HFmap f) a)

Example

suppose we have a function that should be applied to every element of a record:

>>> let circSucc_ x | x == maxBound = minBound | otherwise = succ x

>>> :t circSucc_
circSucc_ :: (Eq a, Enum a, Bounded a) => a -> a

Use a shortcut (Fun) to create a value that has an appropriate ApplyAB instance:

>>> let circSucc = Fun circSucc_ :: Fun '[Eq,Enum,Bounded] '()

Confirm that we got Fun right:

>>> :t applyAB circSucc
applyAB circSucc :: (Bounded b, Enum b, Eq b) => b -> b

>>> applyAB circSucc True
False

define the actual record:

>>> let r = x .=. 'a' .*. y .=. False .*. emptyRecord
>>> r
Record{x='a',y=False}

>>> hMapR circSucc r
Record{x='b',y=True}

newtype HMapR f

Constructors

HMapR f

cast labels

class Relabeled r where

Iso (Record s) (Record t) (Record a) (Record b), such that relabeled = unlabeled . from unlabeled

in other words, pretend a record has different labels, but the same values.

Methods

relabeled :: forall p f s t a b. (HMapTaggedFn (RecordValuesR s) a, HMapTaggedFn (RecordValuesR b) t, SameLengths `[s, a, t, b]`, RecordValuesR t ~ RecordValuesR b, RecordValuesR s ~ RecordValuesR a, RecordValues b, RecordValues s, Profunctor p, Functor f) => (r a `p` f (r b)) -> r s `p` f (r t)

relabeled' :: (Relabeled r, RecordValues b, RecordValues a, RecordValues s, HMapAux HList TaggedFn (RecordValuesR s) a, HMapAux HList TaggedFn (RecordValuesR s) s, SameLength' * * (RecordValuesR s) a, SameLength' * * b s, SameLength' * * a (RecordValuesR s), SameLength' * * a s, SameLength' * * s b, SameLength' * * s a, Profunctor p, Functor f, (~) [*] (RecordValuesR b) (RecordValuesR s), (~) [*] (RecordValuesR a) (RecordValuesR s)) => p (r a) (f (r b)) -> p (r s) (f (r s))

Iso' (Record s) (Record a)

such that RecordValuesR s ~ RecordValuesR a

Hints for type errors

data DuplicatedLabel l

Property of a proper label set for a record: no duplication of labels, and every element of the list is Tagged label value

data ExtraField l

data FieldNotFound l

Unclassified

Probably internals, that may not be useful

zipTagged :: (MapLabel ts ~ lts, HZip Proxy lts vs tvs) => Proxy ts -> proxy vs -> Proxy tvs

a variation on hZip for Proxy, where the list of labels does not have to include Label (as in ts')

>>> let ts = Proxy :: Proxy ["x","y"]
>>> let ts' = Proxy :: Proxy [Label "x",Label "y"]
>>> let vs = Proxy :: Proxy [Int,Char]

>>> :t zipTagged ts Proxy
zipTagged ts Proxy :: Proxy '[Tagged "x" y, Tagged "y" y1]

>>> :t zipTagged ts vs
zipTagged ts vs :: Proxy '[Tagged "x" Int, Tagged "y" Char]

And and the case when hZip does the same thing:

>>> :t zipTagged ts' vs
zipTagged ts' vs :: Proxy '[Tagged "x" Int, Tagged "y" Char]

>>> :t hZip ts' vs
hZip ts' vs :: Proxy '[Tagged "x" Int, Tagged "y" Char]

class HasField' b l r v | b l r -> v where

Methods

hLookupByLabel' :: Proxy b -> Label l -> HList r -> v

type family DemoteMaybe d v :: *

class HasFieldM1 b l r v | b l r -> v where

Methods

hLookupByLabelM1 :: Proxy b -> Label l -> r -> t -> DemoteMaybe t v

class H2ProjectByLabels ls r rin rout | ls r -> rin rout where

Invariant:

r === rin `disjoint-union` rout
labels rin === ls
    where (rin,rout) = hProjectByLabels ls r

Methods

h2projectByLabels :: proxy ls -> HList r -> (HList rin, HList rout)

class H2ProjectByLabels' b ls r rin rout | b ls r -> rin rout where

Methods

h2projectByLabels' :: Proxy b -> proxy ls -> HList r -> (HList rin, HList rout)

class HLabelSet ls

Relation between HLabelSet and HRLabelSet

instance HLabelSet (LabelsOf ps) => HRLabelSet ps

see also HSet

class HLabelSet' l1 l2 leq r

class (HLabelSet (LabelsOf ps), HAllTaggedLV ps) => HRLabelSet ps

class HAllTaggedLV ps

The Record, Variant, TIP, TIC type constructors only make sense when they are applied to an instance of this class

class (HRearrange3 ls r r', LabelsOf r' ~ ls, SameLength ls r, SameLength r r') => HRearrange ls r r' | ls r -> r', r' -> ls where

Helper class for hRearrange

Methods

hRearrange2 :: proxy ls -> HList r -> HList r'

class HRearrange3 ls r r' | ls r -> r' where

same as HRearrange, except no backwards FD

Methods

hRearrange3 :: proxy ls -> HList r -> HList r'

class HRearrange4 l ls rin rout r' | l ls rin rout -> r' where

Helper class 2 for hRearrange

Methods

hRearrange4 :: proxy l -> Proxy ls -> HList rin -> HList rout -> HList r'

class UnionSymRec' b r1 f2 r2' ru | b r1 f2 r2' -> ru where

Methods

unionSR' :: Proxy b -> Record r1 -> f2 -> Record r2' -> (Record ru, Record ru)

type HFindLabel l ls n = HFind l (UnLabel l (LabelsOf ls)) n

A version of HFind where the ls type variable is a list of Tagged or Label. This is a bit indirect, and ideally LabelsOf could have kind [*] -> [k].

labelLVPair :: Tagged l v -> Label l

Label accessor

newLVPair :: Label l -> v -> Tagged l v

type family UnLabel proxy ls :: [k]

remove the Label type constructor. The proxy argument is supplied to make it easier to fix the kind variable k.

type HMemberLabel l r b = HMember l (UnLabel l (LabelsOf r)) b

data TaggedFn

Constructors

TaggedFn

data ReadComponent

type HMapTaggedFn l r = (HMapCxt HList TaggedFn l r, RecordValuesR r ~ l, RecordValues r)

type HLensCxt x r s t a b = (HasField x (r s) a, HUpdateAtLabel r x b s t, HasField x (r t) b, HUpdateAtLabel r x a t s, SameLength s t, SameLabels s t)

constraints needed to implement HLens

zip

use the more general HZip class instead

class HZipRecord x y xy | x y -> xy, xy -> x y where

Methods

hZipRecord :: Record x -> Record y -> Record xy

hUnzipRecord :: Record xy -> (Record x, Record y)

alternative implementation

hZipRecord2 :: (RecordValues y1, RecordValues y, RecordValues x, HMapAux HList TaggedFn (RecordValuesR x) x, HZipList (RecordValuesR y1) (RecordValuesR y) (RecordValuesR x), SameLength' * * y1 x, SameLength' * * y x, SameLength' * * x y1, SameLength' * * x y, SameLabels [*] [*] x y1, SameLabels [*] [*] x y, HAllTaggedLV x) => Record y1 -> Record y -> Record x

instead of explicit recursion above, we could define HZipRecord in terms of HZipList. While all types are inferred, this implementation is probably slower, so explicit recursion is used in the HZip Record instance.

hUnzipRecord2 :: (RecordValues y, RecordValues x1, RecordValues x, HMapAux HList TaggedFn (RecordValuesR x) x, HMapAux HList TaggedFn (RecordValuesR x1) x1, HZipList (RecordValuesR x) (RecordValuesR x1) (RecordValuesR y), SameLength' * * y x1, SameLength' * * y x, SameLength' * * x1 y, SameLength' * * x y, SameLabels [*] [*] x1 y, SameLabels [*] [*] x y, HAllTaggedLV x1, HAllTaggedLV x) => Record y -> (Record x, Record x1)