Data.HList.Variant

Description

The HList library

See CommonMain#Variant for the public (safe) interface.

The implementation here follows Data.Dynamic, though Typeable is not needed.

See broken/VariantP.hs and broken/VariantOld.hs for different approaches to open sums.

Labels for doctests

>>> import Data.HList.RecordPuns
>>> let x = Label :: Label "x"
>>> let y = Label :: Label "y"
>>> let z = Label :: Label "z"
>>> let _left = Label :: Label "left"
>>> let _right = Label :: Label "right"

>>> :set -XQuasiQuotes -XViewPatterns -XDataKinds

- * Creating Variants

It is necessary to specify the order in which the fields occur, using a data type like

>>> let p = Proxy :: Proxy '[Tagged "left" Char, Tagged "right" Int]

Then this argument can be passed into mkVariant

>>> let v = mkVariant _left 'x' p
>>> let w = mkVariant _right 5  p

>>> :t v
v :: Variant '[Tagged "left" Char, Tagged "right" Int]

>>> :t w
w :: Variant '[Tagged "left" Char, Tagged "right" Int]

>>> [v,w]
[V{left='x'},V{right=5}]

Alternative: a Record as the Proxy

The type of mkVariant also allows using a Record as the proxy. For example:

>>> :{
let p2 = [pun| left right |] where
            left = 'a'
            right = (4::Int)
:}

>>> let v2 = mkVariant _left 'x' p2
>>> let w2 = mkVariant _right 5  p2

>>> :t v2
v2 :: Variant '[Tagged "left" Char, Tagged "right" Int]

>>> :t w2
w2 :: Variant '[Tagged "left" Char, Tagged "right" Int]

>>> (v2,w2)
(V{left='x'},V{right=5})

A polymorphic Proxy

It is also possible to leave the Char and Int as type variables, and have them inferred.

>>> let p3 = Proxy :: Proxy '[Tagged "left" a, Tagged "right" b]

Using p3 takes some care. The following attempt shows the problem:

>>> :{
let v3' = mkVariant _left 'x' p3
    w3' = mkVariant _right (5::Int) p3
:}

>>> :t v3'
v3' :: Variant '[Tagged "left" Char, Tagged "right" b]

>>> :t w3'
w3' :: Variant '[Tagged "left" a, Tagged "right" Int]

Here each use of p3 does not constrain the type of the other use. In some cases those type variables will be inferred from other constraints, such as when putting the variants into a list

>>> [v3', w3']
[V{left='x'},V{right=5}]

In other cases the other tags will be defaulted to (), at least if ExtendedDefaultRules is enabled:

>>> v3'
V{left='x'}

>>> :set -XNoExtendedDefaultRules
>>> v3'
...
...No instance for (Show ...) arising from a use of ‘print’
...

Another way around this issue is to make sure that the proxy is bound in a monomorphic pattern. These are patterns that allow name shadowing.

p -> ...

case e of p -> ...

do p <- e; ...

implicit parameters let ?p = e in ...

other patterns involved in mutually recursive bindings

An example of the case:

>>> :{
let (v3,w3) = case p3 of
              p -> (mkVariant _left 'x' p,
                    mkVariant _right (5 :: Int) p)
:}

>>> :t v3
v3 :: Variant '[Tagged "left" Char, Tagged "right" Int]

>>> :t w3
w3 :: Variant '[Tagged "left" Char, Tagged "right" Int]

data Variant vs

Variant vs has an implementation similar to Dynamic, except the contained value is one of the elements of the vs list, rather than being one particular instance of Typeable.

>>> v .!. _right
Nothing

>>> v .!. _left
Just 'x'

In some cases the pun quasiquote works with variants,

>>> let f [pun| left right |] = (left,right)
>>> f v
(Just 'x',Nothing)

>>> f w
(Nothing,Just 5)

>>> let add1 v = hMapV (Fun succ :: Fun '[Enum] '()) v

>>> f (add1 v)
(Just 'y',Nothing)

>>> f (add1 w)
(Nothing,Just 6)

Constructors

Variant !Int Any

Unsafe operations

unsafeMkVariant

Arguments

:: Int
n

-> v

-> Variant vs

This is only safe if the n'th element of vs has type Tagged t v

unsafeCastVariant :: Variant v -> Variant v'

Safe when (e ~ e') given that

Tagged t e ~ HLookupByHNatR n v
Tagged t' e' ~ HLookupByHNatR n v'

hUpdateAtLabel is the safe version

castVariant :: (RecordValuesR v ~ RecordValuesR v', SameLength v v') => Variant v -> Variant v'

in ghc>=7.8, coerce is probably a better choice

unsafeUnVariant :: Variant v -> e

private destructor. This is safe only if the value contained actually has type e

unsafeEmptyVariant :: Variant []

This function is unsafe because it can lead to a runtime error when used together with the HExtend instance (.*.)

>>> print $ (Label :: Label "x") .=. (Nothing :: Maybe ()) .*. unsafeEmptyVariant
V{*** Exception: invalid variant

use mkVariant1 instead

Public constructor

class HasField x (Variant vs) (Maybe v) => MkVariant x v vs | x vs -> v where

Methods

mkVariant

Arguments

:: Label x
the tag

-> v
value to be stored

-> proxy vs
a helper to fix the ordering and types of the potential values that this variant contains. Typically this will be a Proxy, Record or another Variant

-> Variant vs

mkVariant1 :: Label k l -> e -> Variant ((:) * (Tagged k l e) ([] *))

Public destructor

hLookupByLabel (synonym .!.)

(.!.)             :: Variant v -> Label x -> Maybe e
hLookupByLabel    :: Label x -> Variant v -> Maybe e

hPrism and hLens' combine this with mkVariant

splitVariant1 :: Variant (Tagged s x : xs) -> Either x (Variant xs)

splitVariant1' :: Variant (x : xs) -> Either x (Variant xs)

x ~ Tagged s t

extendVariant :: Variant l -> Variant (e : l)

Prism

class (SameLength s t, SameLabels s t) => HPrism x s t a b | x s -> a, x t -> b, x s b -> t, x t a -> s where

Make a Prism (Variant s) (Variant t) a b out of a Label.

See Data.HList.Labelable.hLens' is a more overloaded version.

Few type annotations are necessary because of the restriction that s and t have the same labels in the same order, and to get "t" the "a" in "s" is replaced with "b".

Methods

hPrism :: (Choice p, Applicative f) => Label x -> p a (f b) -> p (Variant s) (f (Variant t))

Read

class ShowVariant vs where

helper class for defining the Show instance

Methods

showVariant :: Variant vs -> ShowS

Show

class ReadVariant vs where

Methods

readVariant :: ReadP (Variant vs)

Data

class VariantConstrs xs where

Methods

variantConstrs :: DataType -> proxy xs -> [Constr]

class GunfoldVariant es v where

implementation of gunfold for Variant

In ghci

:set -ddump-deriv -XDeriveDataTypeable
data X a b c = A a | B b | C c deriving (Data,Typeable)

shows that gunfold is defined something like

gunfold k z c = case constrIndex c of
      1 -> k (z Ghci1.A)
      2 -> k (z Ghci1.B)
      _ -> k (z Ghci1.C)

If we instead had

type X a b c = Variant [Tagged "A" a, Tagged "B" b, Tagged "C" c]

Then we could write:

gunfold1 :: (forall b r. Data b => (b -> r) -> c r)
         -> Variant [Tagged "A" a, Tagged "B" b, Tagged "C" c]
gunfold1 f c = case constrIndex c of
      1 -> f mkA
      2 -> f mkB
      _ -> f mkC
  where mkA a = mkVariant (Label :: Label "A") (a :: a) v
        mkB b = mkVariant (Label :: Label "B") (b :: b) v
        mkC c = mkVariant (Label :: Label "C") (c :: c) v
        v = Proxy :: Proxy [Tagged "A" a, Tagged "B" b, Tagged "C" c]

where f = k.z

Methods

gunfoldVariant

Arguments

:: (forall b. Data b => (b -> Variant v) -> c (Variant v))
f = k . z

-> Proxy es

-> Int

-> c (Variant v)

class GfoldlVariant xs xs' where

Methods

gfoldlVariant :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> Variant xs -> c (Variant xs')

the same as gfoldl, except the variant that is returned can have more possible values (needed to actually implement gfoldl).

Map

newtype HMapV f

Apply a function to all possible elements of the variant

Constructors

HMapV f

hMapV :: (HMapAux Variant (HFmap f) x y, SameLength' * * y x, SameLength' * * x y) => f -> Variant x -> Variant y

shortcut for applyAB . HMapV. hMap is more general

hMapOutV :: forall x y z f. (SameLength x y, HMapAux Variant (HFmap f) x y, Unvariant y z, HMapOutV_gety x z ~ y) => f -> Variant x -> z

hMapOutV f = unvariant . hMapV f, except an ambiguous type variable is resolved by HMapOutV_gety

type family HMapOutV_gety x z :: [*]

resolves an ambiguous type in hMapOutV

HUpdateAtLabel instance

type HUpdateVariantAtLabelCxt l e v v' n _e = (HFindLabel l v n, HFindLabel l v' n, HUpdateAtHNatR n (Tagged l e) v ~ v', HasField l (Variant v) (Maybe _e), HasField l (Record v') e, MkVariant l e v')

HExtend instance

type family UnMaybe le

Conversion to an untagged value

class HAllEqVal x b | x -> b

class HAllEqVal' x

class Unvariant' v e | v -> e where

Similar to unvariant, except type variables in v will be made equal to e if possible. That allows the type of Nothing to be inferred as Maybe Char.

>>> unvariant' $ x .=. Nothing .*. mkVariant1 y 'y'
'y'

However, this difference leads to more local error messages (Couldn't match type ‘()’ with ‘Char’), rather than the following with unvariant:

Fail
   '("Variant",
     '[Tagged "left" Char, Tagged "right" ()],
     "must have all values equal to ",
     e))

Methods

unvariant' :: Variant v -> e

class Unvariant v e | v -> e where

Convert a Variant which has all possibilities having the same type into a value of that type. Analogous to either id id.

See also unvariant'

Methods

unvariant :: Variant v -> e

class Unvariant1 b v e | b v -> e where

Methods

unvariant1 :: Proxy b -> Variant v -> e

data UnvariantTypeMismatch vs

unvarianted :: (Unvariant' s a, Unvariant' t b, SameLabels s t, SameLength s t, Functor f) => (a -> f b) -> Variant s -> f (Variant t)

Lens (Variant s) (Variant t) a b

Analogue of Control.Lens.chosen :: Lens (Either a a) (Either b b) a b

unvarianted' :: (HAllEqVal ((:) * (Tagged * () b) s) b1, HAllEqVal s b1, HAllEqVal' ((:) * (Tagged * () b) s), Unvariant1 Bool b1 s b, SameLength' * * s s, SameLabels [*] [*] s s, Functor f) => (b -> f b) -> Variant s -> f (Variant s)

Lens' (Variant s) a

where we might have s ~ '[Tagged t1 a, Tagged t2 a]

Zip

class ZipVariant x y xy | x y -> xy, xy -> x y where

Applies to variants that have the same labels in the same order. A generalization of

zipEither :: Either a b -> Either a b -> Maybe (Either (a,a) (b,b))
zipEither (Left a) (Left a') = Just (Left (a,a'))
zipEither (Right a) (Right a') = Just (Right (a,a'))
zipEither _ _ = Nothing

see HZip for zipping other collections

Methods

zipVariant :: Variant x -> Variant y -> Maybe (Variant xy)

with a record

class (SameLength v v', SameLabels v v') => ZipVR fs v v' | fs v -> v' where

Apply a record of functions to a variant of values. The functions are selected based on those having the same label as the value.

Methods

zipVR_ :: Record fs -> Variant v -> Variant v'

zipVR is probably a better choice in most situations, since it requires that fs has one function for every element of v

zipVR :: (SameLabels fs v, SameLength fs v, ZipVR fs v v', ZipVRCxt fs v v') => Record fs -> Variant v -> Variant v'

>>> let xy = x .*. y .*. emptyProxy
>>> let p = Proxy `asLabelsOf` xy
>>> let vs = [ mkVariant x 1.0 p, mkVariant y () p ]

>>> zipVR (hBuild (+1) id) `map` vs
[V{x=2.0},V{y=()}]

type family ZipVRCxt fs xs ys :: Constraint

Lets zipVR act as if ZipVR fs v v' had an FD v v' -> fs

ZipVRCxt [Tagged s f,  Tagged t g]
         [Tagged s fx, Tagged t gx]
         [Tagged s fy, Tagged t gy]
  = (f ~ (fx -> fy), g ~ (gx -> gy))

Eq

Alternative Eq

eqVariant :: (HAllEqVal ((:) * (Tagged * () Bool) (HMapOutV_gety x Bool)) b, HAllEqVal (HMapOutV_gety x Bool) b, Unvariant1 Bool b (HMapOutV_gety x Bool) Bool, ZipVariant x1 y x, HMapAux Variant (HFmap UncurryEq) x (HMapOutV_gety x Bool), SameLength' * * (HMapOutV_gety x Bool) x, SameLength' * * x (HMapOutV_gety x Bool)) => Variant x1 -> Variant y -> Bool

implemented like and (zipWith (==) xs ys). Behaves the same as the Eq instances for Variant

data UncurryEq

Constructors

UncurryEq

Ord

Bounded

Enum

Ix (TODO)

Monoid

Projection

class ProjectVariant x y where

convert a variant with more fields into one with fewer (or the same) fields.

>>> let ty = Proxy :: Proxy [Tagged "left" Int, Tagged "right" Int]
>>> let l = mkVariant _left 1 ty
>>> let r = mkVariant _right 2 ty

>>> map projectVariant [l, r] :: [Maybe (Variant '[Tagged "left" Int])]
[Just V{left=1},Nothing]

rearrangeVariant = fromJust . projectVariant is one implementation of rearrangeVariant, since the result can have the same fields with a different order:

>>> let yt = Proxy :: Proxy [Tagged "right" Int, Tagged "left" Int]

>>> map projectVariant [l, r] `asTypeOf` [Just (mkVariant _left 0 yt)]
[Just V{left=1},Just V{right=2}]

Methods

projectVariant :: Variant x -> Maybe (Variant y)

class HAllTaggedLV y => ProjectExtendVariant x y where

projectExtendVariant = fmap extendVariant . projectVariant

where intermediate variant is as large as possible. Used to implement Data.HList.Labelable.projected

Note that:

>>> let r = projectExtendVariant (mkVariant1 Label 1 :: Variant '[Tagged "x" Int])
>>> r :: Maybe (Variant '[Tagged "x" Integer])
Nothing

Methods

projectExtendVariant :: Variant x -> Maybe (Variant y)

class ProjectExtendVariant' inY lv y where

Methods

projectExtendVariant' :: Proxy inY -> lv -> Maybe (Variant y)

class (ProjectVariant x yin, ProjectVariant x yout) => SplitVariant x yin yout where

Methods

splitVariant :: Variant x -> Either (Variant yin) (Variant yout)

class (HAllTaggedLV y, HAllTaggedLV x) => ExtendsVariant x y where

projectVariant . extendsVariant = Just (when the types match up)

extendVariant is a special case

Methods

extendsVariant :: Variant x -> Variant y

rearrangeVariant :: (SameLength v v', ExtendsVariant v v') => Variant v -> Variant v'

rearrangeVariant is a specialization of extendsVariant whose result is always . see also rearranged

hMaybied :: (VariantToHMaybied v1 x, VariantToHMaybied v r, HFoldr HMaybiedToVariantFs [Variant ([] *)] x [Variant v1], HMapAux HList (HFmap HCastF) x r, SameLength' * * x r, SameLength' * * r x, Choice p, Applicative f) => p (Variant v1) (f (Variant v)) -> p (Record x) (f (Record r))

Prism (Record tma) (Record tmb) (Variant ta) (Variant tb)

see hMaybied'

data HCastF

Constructors

HCastF

hMaybied' :: (VariantToHMaybied v x, HFoldr HMaybiedToVariantFs [Variant ([] *)] x [Variant v], HMapAux HList (HFmap HCastF) x x, SameLength' * * x x, Choice p, Applicative f) => p (Variant v) (f (Variant v)) -> p (Record x) (f (Record x))

Prism' (Record tma) (Variant ta)

where tma and tmb are lists like

tma ~ '[Tagged x (Maybe a), Tagged y (Maybe b)]
ta  ~ '[Tagged x        a , Tagged y        b ]

If one element of the record is Just, the Variant will contain that element. Otherwise, the prism fails.

Note

The types work out to define a prism:

l = prism' variantToHMaybied (listToMaybe . hMaybiedToVariants)

but the law: s^?l ≡ Just a ==> l # a ≡ s is not followed, because we could have:

  s, s2 :: Record '[Tagged "x" (Maybe Int), Tagged "y" (Maybe Char)]
  s = hBuild (Just 1) (Just '2')
  s2 = hBuild (Just 1) Nothing

  v :: Variant '[Tagged "x" Int, Tagged "y" Char]
  v = mkVariant (Label :: Label "x") 1 Proxy

So that s^?l == Just v. But l#v == s2 /= s, while the law requires l#v == s. hMaybied avoids this problem by only producing a value when there is only one present.

class VariantToHMaybied v r | v -> r, r -> v where

Methods

variantToHMaybied :: Variant v -> Record r

data ConstTaggedNothing

Constructors

ConstTaggedNothing

hMaybiedToVariants :: (HFoldr HMaybiedToVariantFs [Variant []] r [Variant v], VariantToHMaybied v r) => Record r -> [Variant v]

Every element of the record that is Just becomes one element in the resulting list. See hMaybied' example types that r and v can take.

data HMaybiedToVariantFs

Constructors

HMaybiedToVariantFs